Carve parser production code into themed submodules
parser/mod.rs keeps the types and the parse_line driver; the method
bodies move verbatim (indentation included) into core.rs (executor
token loop), speculation.rs (branch-point fail handling and corrected-
op arbitration), semantics.rs (push_meta_ops/perform_op/capture ops),
and embed.rs (escape execution). The five methods called across
submodule boundaries become pub(super); no other signature changes.
Tests remain inline in mod.rs for the follow-up split.
