+++ src/parsing/parser/mod.rs
@@ -336,137 +336,10 @@ struct RegexMatch<'a> {
 /// Maps the pattern to the start index, which is -1 if not found.
 type SearchCache = HashMap<*const MatchPattern, Option<Region>, BuildHasherDefault<FnvHasher>>;
 
-/// Build the ordered Push/Pop ops for a match's `captures:` mapping over
-/// its regex `regions`. Captures can appear in arbitrary source order
-/// (e.g. `((bob)|(hi))*` matching `hibob` — the outer group must Push
-/// before any inner group). Empty captures are skipped because they'd
-/// otherwise sort a Pop before its Push. The returned ops are already
-/// position-ordered and safe to append to a parser ops vec.
-///
-/// Each capture's `(cap_start, cap_end)` is clipped to the outer match
-/// range `regions.pos(0)` so a group matching inside a `(?=...)` /
-/// `(?<=...)` whose own span extends past the consumed range still
-/// colours the overlap (the boundary char) and nothing beyond it. This
-/// mirrors Sublime Text; without the clip, a lookahead-internal
-/// `captures:` entry leaked scope over unmatched trailing chars or was
-/// silently dropped upstream in `parse_captures`.
-fn build_capture_ops(capture_map: &CaptureMapping, regions: &Region) -> Vec<(usize, ScopeStackOp)> {
-    let mut map: Vec<((usize, i32), ScopeStackOp)> = Vec::new();
-    let (match_start, match_end) = match regions.pos(0) {
-        Some(bounds) => bounds,
-        None => return Vec::new(),
-    };
-    for &(cap_index, ref scopes) in capture_map.iter() {
-        if let Some((cap_start, cap_end)) = regions.pos(cap_index) {
-            let clipped_start = cap_start.max(match_start);
-            let clipped_end = cap_end.min(match_end);
-            if clipped_start >= clipped_end {
-                continue;
-            }
-            for scope in scopes.iter() {
-                map.push((
-                    (clipped_start, -((clipped_end - clipped_start) as i32)),
-                    ScopeStackOp::Push(*scope),
-                ));
-            }
-            map.push(((clipped_end, i32::MIN), ScopeStackOp::Pop(scopes.len())));
-        }
-    }
-    map.sort_by(|a, b| a.0.cmp(&b.0));
-    map.into_iter().map(|((i, _), op)| (i, op)).collect()
-}
-
-// To understand the implementation of this, here's an introduction to how
-// Sublime Text syntax definitions work.
-//
-// Let's say we have the following made-up syntax definition:
-//
-//     contexts:
-//       main:
-//         - match: A
-//           scope: scope.a.first
-//           push: context-a
-//         - match: b
-//           scope: scope.b
-//         - match: \w+
-//           scope: scope.other
-//       context-a:
-//         - match: a+
-//           scope: scope.a.rest
-//         - match: (?=.)
-//           pop: true
-//
-// There are two contexts, `main` and `context-a`. Each context contains a list
-// of match rules with instructions for how to proceed.
-//
-// Let's say we have the input string " Aaaabxxx". We start at position 0 in
-// the string. We keep a stack of contexts, which at the beginning is just main.
-//
-// So we start by looking at the top of the context stack (main), and look at
-// the rules in order. The rule that wins is the first one that matches
-// "earliest" in the input string. In our example:
-//
-// 1. The first one matches "A". Note that matches are not anchored, so this
-//    matches at position 1.
-// 2. The second one matches "b", so position 5. The first rule is winning.
-// 3. The third one matches "\w+", so also position 1. But because the first
-//    rule comes first, it wins.
-//
-// So now we execute the winning rule. Whenever we matched some text, we assign
-// the scope (if there is one) to the matched text and advance our position to
-// after the matched text. The scope is "scope.a.first" and our new position is
-// after the "A", so 2. The "push" means that we should change our stack by
-// pushing `context-a` on top of it.
-//
-// In the next step, we repeat the above, but now with the rules in `context-a`.
-// The result is that we match "a+" and assign "scope.a.rest" to "aaa", and our
-// new position is now after the "aaa". Note that there was no instruction for
-// changing the stack, so we stay in that context.
-//
-// In the next step, the first rule doesn't match anymore, so we go to the next
-// rule where "(?=.)" matches. The instruction is to "pop", which means we
-// pop the top of our context stack, which means we're now back in main.
-//
-// This time in main, we match "b", and in the next step we match the rest with
-// "\w+", and we're done.
-//
-//
-// ## Preventing loops
-//
-// These are the basics of how matching works. Now, you saw that you can write
-// patterns that result in an empty match and don't change the position. These
-// are called non-consuming matches. The problem with them is that they could
-// result in infinite loops. Let's look at a syntax where that is the case:
-//
-//     contexts:
-//       main:
-//         - match: (?=.)
-//           push: test
-//       test:
-//         - match: \w+
-//           scope: word
-//         - match: (?=.)
-//           pop: true
-//
-// This is a bit silly, but it's a minimal example for explaining how matching
-// works in that case.
-//
-// Let's say we have the input string " hello". In `main`, our rule matches and
-// we go into `test` and stay at position 0. Now, the best match is the rule
-// with "pop". But if we used that rule, we'd pop back to `main` and would still
-// be at the same position we started at! So this would be an infinite loop,
-// which we don't want.
-//
-// So what Sublime Text does in case a looping rule "won":
-//
-// * If there's another rule that matches at the same position and does not
-//   result in a loop, use that instead.
-// * Otherwise, go to the next position and go through all the rules in the
-//   current context again. Note that it means that the "pop" could again be the
-//   winning rule, but that's ok as it wouldn't result in a loop anymore.
-//
-// So in our input string, we'd skip one character and try to match the rules
-// again. This time, the "\w+" wins because it comes first.
+mod core;
+mod embed;
+mod semantics;
+mod speculation;
 
 impl ParseState {
     /// Creates a state from a syntax definition, keeping its own reference-counted point to the
@@ -691,3031 +564,6 @@ impl ParseState {
 
         Ok(res)
     }
-
-    #[allow(clippy::too_many_arguments)]
-    fn parse_next_token(
-        &mut self,
-        line: &str,
-        syntax_set: &SyntaxSet,
-        start: &mut usize,
-        search_cache: &mut SearchCache,
-        regions: &mut Region,
-        non_consuming_push_at: &mut (usize, usize, usize),
-        ops: &mut Vec<(usize, ScopeStackOp)>,
-    ) -> Result<bool, ParsingError> {
-        let (check_pop_loop, pre_push_depth) = {
-            let (pos, pre, post) = *non_consuming_push_at;
-            let armed = pos == *start && pre < self.stack.len() && self.stack.len() <= post;
-            (armed, pre)
-        };
-
-        // Trim proto_starts that are no longer valid
-        while self
-            .proto_starts
-            .last()
-            .map(|start| *start >= self.stack.len())
-            .unwrap_or(false)
-        {
-            self.proto_starts.pop();
-        }
-
-        let best_match = self.find_best_match(
-            line,
-            *start,
-            syntax_set,
-            search_cache,
-            regions,
-            check_pop_loop,
-            pre_push_depth,
-        )?;
-
-        if let Some(reg_match) = best_match {
-            // Check if this is an escape match (sentinel pat_index)
-            if reg_match.pat_index == usize::MAX {
-                let (match_start, match_end) = reg_match.regions.pos(0).unwrap();
-                let zero_width = match_start == match_end;
-
-                if zero_width
-                    && self
-                        .zero_width_escape_fires
-                        .get(&match_start)
-                        .map(|n| *n >= ZERO_WIDTH_ESCAPE_FIRE_LIMIT)
-                        .unwrap_or(false)
-                {
-                    // Zero-width escape repeatedly firing at the same byte
-                    // within a single `parse_line_inner_from` — diagnostic of
-                    // a branch-fail rewind cycle (#650). Suppress and advance
-                    // one character to break the loop, mirroring the
-                    // `would_loop` enforcement below. Threshold lets
-                    // legitimate alt-cycle replays through `handle_fail`
-                    // re-encounter the same offset without false-tripping.
-                    if let Some((i, _)) = line[*start..].char_indices().nth(1) {
-                        *start += i;
-                        return Ok(true);
-                    } else {
-                        return Ok(false);
-                    }
-                }
-
-                *start = match_end;
-                self.exec_escape(
-                    reg_match.escape_index,
-                    match_start,
-                    match_end,
-                    &reg_match.regions,
-                    syntax_set,
-                    ops,
-                )?;
-                if zero_width {
-                    *self.zero_width_escape_fires.entry(match_start).or_insert(0) += 1;
-                }
-                search_cache.clear();
-                return Ok(true);
-            }
-
-            if reg_match.would_loop {
-                // A push that doesn't consume anything (a regex that resulted
-                // in an empty match at the current position) can not be
-                // followed by a non-consuming pop. Otherwise we're back where
-                // we started and would try the same sequence of matches again,
-                // resulting in an infinite loop. In this case, Sublime Text
-                // advances one character and tries again, thus preventing the
-                // loop.
-
-                // println!("pop_would_loop for match {:?}, start {}", reg_match, *start);
-
-                // nth(1) gets the next character if there is one. Need to do
-                // this instead of just += 1 because we have byte indices and
-                // unicode characters can be more than 1 byte.
-                if let Some((i, _)) = line[*start..].char_indices().nth(1) {
-                    *start += i;
-                    return Ok(true);
-                } else {
-                    // End of line, no character to advance and no point trying
-                    // any more patterns.
-                    return Ok(false);
-                }
-            }
-
-            let match_end = reg_match.regions.pos(0).unwrap().1;
-
-            // Check if this is a Fail operation — handle before advancing start
-            let context = reg_match.context;
-            let match_pattern = context.match_at(reg_match.pat_index)?;
-            if let MatchOperation::Fail(_) = match_pattern.operation {
-                let level_context = {
-                    let id = &self.stack[self.stack.len() - 1].context;
-                    syntax_set.get_context(id)?
-                };
-                return self.exec_pattern(
-                    line,
-                    &reg_match,
-                    level_context,
-                    syntax_set,
-                    start,
-                    non_consuming_push_at,
-                    ops,
-                    search_cache,
-                );
-            }
-
-            let consuming = match_end > *start;
-            if !consuming {
-                // The match doesn't consume any characters. If this is a
-                // "push", remember the position and the post-push stack depth
-                // interval so that we can check the next "pop" for loops.
-                // Otherwise leave the state, e.g. non-consuming "set" could
-                // also result in a loop.
-                //
-                // The interval [pre+1, pre+K] covers every depth the unwind
-                // can pass through; for K=1 it collapses to the original
-                // armed depth pre+1, preserving the prior single-push guard.
-                let k = match &match_pattern.operation {
-                    MatchOperation::Push { ctx_refs, .. } => ctx_refs.len(),
-                    MatchOperation::Branch { .. } => 1,
-                    MatchOperation::Embed { contexts, .. } => contexts.len(),
-                    _ => 0,
-                };
-                if k > 0 {
-                    let pre = self.stack.len();
-                    let post = pre + k;
-                    *non_consuming_push_at = (match_end, pre, post);
-                }
-                // Inside a cross-line replay, a non-consuming `Branch` whose
-                // match lands past every character of the replay line creates
-                // a chained `branch_point` that the outer parse will then
-                // exhaust on the next (often empty) line. The exhaustion's
-                // pop ops attach to *this* replay line — i.e. the next line's
-                // baseline — collapsing the parent context one boundary too
-                // early. Skip the branch creation; the parent rule will fire
-                // again at the start of the outer line and the BP will be
-                // anchored there instead. Observed on Markdown's LRD blank
-                // line: link-def-attr's `match: $` was creating a BP-2
-                // inside link-title-continuation's exhaustion replay, then
-                // collapsing `meta.link.reference.def.markdown` on the empty
-                // line.
-                if matches!(match_pattern.operation, MatchOperation::Branch { .. })
-                    && self.replay_ctx.is_some()
-                    && match_end >= line.len()
-                {
-                    return Ok(false);
-                }
-            }
-
-            *start = match_end;
-
-            // Prune stale skipped_branches entries: a BP marked as skipped
-            // at an older cursor position no longer matters once the
-            // cursor has advanced past that position.
-            self.skipped_branches.retain(|(c, _)| *c >= *start);
-
-            // ignore `with_prototype`s below this if a context is pushed
-            if reg_match.from_with_prototype {
-                // use current height, since we're before the actual push
-                self.proto_starts.push(self.stack.len());
-            }
-
-            let level_context = {
-                let id = &self.stack[self.stack.len() - 1].context;
-                syntax_set.get_context(id)?
-            };
-            self.exec_pattern(
-                line,
-                &reg_match,
-                level_context,
-                syntax_set,
-                start,
-                non_consuming_push_at,
-                ops,
-                search_cache,
-            )?;
-
-            Ok(true)
-        } else if self.skipped_branches.iter().any(|(c, _)| *c == *start) {
-            // No pattern matched, but we suppressed at least one Branch
-            // at this cursor (its alts all exhausted). Advance one char as
-            // a last resort to break the loop.
-            self.skipped_branches.retain(|(c, _)| *c != *start);
-            if let Some((i, _)) = line[*start..].char_indices().nth(1) {
-                *start += i;
-                search_cache.clear();
-                Ok(true)
-            } else {
-                Ok(false)
-            }
-        } else {
-            Ok(false)
-        }
-    }
-
-    fn find_best_match<'a>(
-        &self,
-        line: &str,
-        start: usize,
-        syntax_set: &'a SyntaxSet,
-        search_cache: &mut SearchCache,
-        regions: &mut Region,
-        check_pop_loop: bool,
-        pre_push_depth: usize,
-    ) -> Result<Option<RegexMatch<'a>>, ParsingError> {
-        let cur_level = &self.stack[self.stack.len() - 1];
-        let context = syntax_set.get_context(&cur_level.context)?;
-        let prototype = if let Some(ref p) = context.prototype {
-            Some(p)
-        } else {
-            None
-        };
-
-        // Build an iterator for the contexts we want to visit in order
-        let context_chain = {
-            let proto_start = self.proto_starts.last().cloned().unwrap_or(0);
-            // Sublime applies with_prototypes from bottom to top
-            let with_prototypes = self.stack[proto_start..].iter().flat_map(|lvl| {
-                lvl.prototypes
-                    .iter()
-                    .map(move |ctx| (true, ctx, lvl.captures.as_ref()))
-            });
-            let cur_prototype = prototype.into_iter().map(|ctx| (false, ctx, None));
-            let cur_context =
-                Some((false, &cur_level.context, cur_level.captures.as_ref())).into_iter();
-            with_prototypes.chain(cur_prototype).chain(cur_context)
-        };
-
-        // println!("{:#?}", cur_level);
-        // println!("token at {} on {}", start, line.trim_right());
-
-        // Check escape patterns first — they take strict precedence.
-        // If an escape matches at `start`, return it immediately as a synthetic match.
-        // If it matches later, truncate the search region for normal patterns.
-        let mut search_end = line.len();
-        let mut escape_match: Option<(usize, Region)> = None; // (escape_stack_index, region)
-
-        for (ei, entry) in self.escape_stack.iter().enumerate() {
-            let mut esc_regions = Region::new();
-            if entry
-                .regex
-                .search(line, start, line.len(), Some(&mut esc_regions), true)
-            {
-                let (esc_start, _esc_end) = esc_regions.pos(0).unwrap();
-                if esc_start < search_end {
-                    search_end = esc_start;
-                    escape_match = Some((ei, esc_regions));
-                }
-            }
-        }
-
-        // If escape matches right at `start`, it wins immediately — no need to
-        // search normal patterns.
-        if let Some((ei, ref esc_region)) = escape_match {
-            let esc_start = esc_region.pos(0).unwrap().0;
-            if esc_start == start {
-                return Ok(Some(RegexMatch {
-                    regions: esc_region.clone(),
-                    context: syntax_set.get_context(&cur_level.context)?,
-                    pat_index: usize::MAX, // sentinel for escape match
-                    from_with_prototype: false,
-                    would_loop: false,
-                    escape_index: ei,
-                }));
-            }
-        }
-
-        let mut min_start = usize::MAX;
-        let mut best_match: Option<RegexMatch<'_>> = None;
-        let mut pop_would_loop = false;
-
-        for (from_with_proto, ctx, captures) in context_chain {
-            for (pat_context, pat_index) in context_iter(syntax_set, syntax_set.get_context(ctx)?) {
-                let match_pat = pat_context.match_at(pat_index)?;
-
-                // Skip Branch patterns whose name was just exhausted at this
-                // cursor. See ParseState::skipped_branches and the same-line
-                // exhaustion handler in handle_fail.
-                if let MatchOperation::Branch { name, .. } = &match_pat.operation {
-                    if self
-                        .skipped_branches
-                        .iter()
-                        .any(|(c, n)| *c == start && n == name)
-                    {
-                        continue;
-                    }
-                }
-
-                if let Some(match_region) = self.search_with_end(
-                    line,
-                    start,
-                    search_end,
-                    match_pat,
-                    captures,
-                    search_cache,
-                    regions,
-                ) {
-                    let (match_start, match_end) = match_region.pos(0).unwrap();
-
-                    // println!("matched pattern {:?} at start {} end {} (pop would loop: {}, min start: {}, initial start: {}, check_pop_loop: {}, stack_len: {})", match_pat, match_start, match_end, pop_would_loop, min_start, start, check_pop_loop, self.stack.len());
-
-                    if match_start < min_start || (match_start == min_start && pop_would_loop) {
-                        // New match is earlier in text than old match,
-                        // or old match was a looping pop at the same
-                        // position.
-
-                        // println!("setting as current match");
-
-                        min_start = match_start;
-
-                        let consuming = match_end > start;
-                        // A non-consuming `pop: N` after a non-consuming push
-                        // loops iff its post-pop depth equals the trigger
-                        // depth — strict equality. Dropping *below*
-                        // pre_push_depth leaves the trigger context entirely
-                        // (e.g. Haskell's `immediately-pop2` as the fallback
-                        // branch alternative for `declaration-type-end`).
-                        // Generalises the prior `Pop(1)` narrowing (which
-                        // matched K=1) to multi-context pushes where the loop
-                        // closes via a multi-level pop chain (e.g. `pop:1`
-                        // then `pop:2` against `push:[a,b,c]`).
-                        pop_would_loop = check_pop_loop
-                            && !consuming
-                            && match &match_pat.operation {
-                                MatchOperation::Pop(n) => {
-                                    self.stack.len().saturating_sub(*n) == pre_push_depth
-                                }
-                                _ => false,
-                            };
-
-                        let push_too_deep = matches!(
-                            match_pat.operation,
-                            MatchOperation::Push { .. }
-                                | MatchOperation::Branch { .. }
-                                | MatchOperation::Embed { .. }
-                        ) && self.stack.len() >= 100;
-
-                        if push_too_deep {
-                            return Ok(None);
-                        }
-
-                        best_match = Some(RegexMatch {
-                            regions: match_region,
-                            context: pat_context,
-                            pat_index,
-                            from_with_prototype: from_with_proto,
-                            would_loop: pop_would_loop,
-                            escape_index: 0, // not an escape match
-                        });
-
-                        if match_start == start && !pop_would_loop {
-                            // We're not gonna find a better match after this,
-                            // so as an optimization we can stop matching now.
-                            return Ok(best_match);
-                        }
-                    }
-                }
-            }
-        }
-
-        // If no normal match was found before the escape position, or escape
-        // position is earlier, use the escape match.
-        if let Some((ei, esc_region)) = escape_match {
-            let esc_start = esc_region.pos(0).unwrap().0;
-            if esc_start < min_start || (esc_start == min_start && pop_would_loop) {
-                return Ok(Some(RegexMatch {
-                    regions: esc_region,
-                    context: syntax_set.get_context(&cur_level.context)?,
-                    pat_index: usize::MAX, // sentinel for escape match
-                    from_with_prototype: false,
-                    would_loop: false,
-                    escape_index: ei,
-                }));
-            }
-        }
-
-        Ok(best_match)
-    }
-
-    fn search_with_end(
-        &self,
-        line: &str,
-        start: usize,
-        search_end: usize,
-        match_pat: &MatchPattern,
-        captures: Option<&(Region, String)>,
-        search_cache: &mut SearchCache,
-        regions: &mut Region,
-    ) -> Option<Region> {
-        // println!("{} - {:?} - {:?}", match_pat.regex_str, match_pat.has_captures, cur_level.captures.is_some());
-        let match_ptr = match_pat as *const MatchPattern;
-
-        // Only consult the cache when searching the full line. Cached entries
-        // are produced under full-line lookahead semantics: a truncated search
-        // at an embed-escape boundary may flip lookahead/lookbehind results
-        // that depended on chars past `search_end`. Concretely, ``done`` at
-        // the close of a backticked `for…done` would be cached as
-        // no-match (the keyword's `(?!cmd_char)` saw the closing backtick
-        // through full-line search) and then short-circuited inside the
-        // backtick embed even though the lookahead actually succeeds against
-        // the embed's escape boundary.
-        if search_end == line.len() {
-            if let Some(maybe_region) = search_cache.get(&match_ptr) {
-                if let Some(ref region) = *maybe_region {
-                    let (cached_start, _cached_end) = region.pos(0).unwrap();
-                    if cached_start >= start {
-                        return Some(region.clone());
-                    }
-                    // cached_start < start: cache miss, re-search below
-                } else {
-                    // Didn't find a match earlier, so no point trying again.
-                    return None;
-                }
-            }
-        }
-
-        let (regex, can_cache) = match (match_pat.has_captures, captures) {
-            (true, Some(captures)) => {
-                let (region, s) = captures;
-                (&match_pat.regex_with_refs(region, s), false)
-            }
-            _ => (match_pat.regex(), true),
-        };
-        // Only `MatchOperation::None` patterns must avoid zero-length matches; every other
-        // operation legitimately needs them (lookaheads with branch/fail, empty patterns with
-        // pop/set, etc.). The regex engine handles this via its `FIND_NOT_EMPTY` option.
-        let allow_empty = !matches!(match_pat.operation, MatchOperation::None);
-        // print!("  executing regex: {:?} at pos {} on line {}", regex.regex_str(), start, line);
-        let matched = regex.search(line, start, search_end, Some(regions), allow_empty);
-
-        if matched {
-            let (match_start, match_end) = regions.pos(0).unwrap();
-            // this is necessary to avoid infinite looping on dumb patterns
-            let does_something = match match_pat.operation {
-                MatchOperation::None => match_start != match_end,
-                MatchOperation::Push { .. }
-                | MatchOperation::Branch { .. }
-                | MatchOperation::Embed { .. } => self.stack.len() < 100,
-                _ => true,
-            };
-            if can_cache && does_something && search_end == line.len() {
-                // Only cache when searching the full line — truncated searches
-                // could give different results for later positions.
-                search_cache.insert(match_pat, Some(regions.clone()));
-            }
-            if does_something {
-                // print!("catch {} at {} on {}", match_pat.regex_str, match_start, line);
-                return Some(regions.clone());
-            }
-        } else if can_cache && search_end == line.len() {
-            search_cache.insert(match_pat, None);
-        }
-        None
-    }
-
-    /// Returns true if the stack was changed.
-    /// For `Fail` operations, returns `Ok(true)` if backtracking was performed
-    /// (caller should continue parsing from the rewound position).
-    fn exec_pattern<'a>(
-        &mut self,
-        line: &str,
-        reg_match: &RegexMatch<'a>,
-        level_context: &'a Context,
-        syntax_set: &'a SyntaxSet,
-        start: &mut usize,
-        non_consuming_push_at: &mut (usize, usize, usize),
-        ops: &mut Vec<(usize, ScopeStackOp)>,
-        search_cache: &mut SearchCache,
-    ) -> Result<bool, ParsingError> {
-        let (match_start, match_end) = reg_match.regions.pos(0).unwrap();
-        let context = reg_match.context;
-        let pat = context.match_at(reg_match.pat_index)?;
-
-        // Handle Fail: attempt backtracking
-        if let MatchOperation::Fail(ref name) = pat.operation {
-            return self.handle_fail(
-                name,
-                line,
-                start,
-                non_consuming_push_at,
-                ops,
-                search_cache,
-                syntax_set,
-            );
-        }
-
-        // For Branch, we need to snapshot state before executing, then synthesize a Push.
-        let is_branch = matches!(pat.operation, MatchOperation::Branch { .. });
-        let synthetic_op;
-
-        if is_branch {
-            if let MatchOperation::Branch {
-                ref name,
-                ref alternatives,
-                pop_count,
-            } = pat.operation
-            {
-                // Snapshot current state.
-                //
-                // NOTE on field naming: `match_start` here stores the
-                // position the parser should *resume* from on fail —
-                // which is the branch match's end position (since the
-                // parser has already consumed the match). `match_end`
-                // and `pat_scope` carry the *real* match span plus the
-                // keyword's own scopes so a same-line fail rewind can
-                // re-emit them (they were truncated off `ops` along
-                // with the alt[0]'s subsequent work).
-                // When `handle_fail` is mid-replay, `self.line_number` /
-                // `self.pending_lines` still reflect the *outer* current
-                // line — read through `replay_ctx` so a branch born
-                // during replay anchors to the virtual replay line `L+i`.
-                let (bp_line_number, bp_pending_lines_snapshot_len) = match &self.replay_ctx {
-                    Some(ctx) => (ctx.line_number, ctx.pending_lines_snapshot_offset),
-                    None => (self.line_number.saturating_sub(1), self.pending_lines.len()),
-                };
-                // When this branch is born inside an outer cross-line
-                // replay's `parse_line_inner_from`, the local `ops` Vec
-                // is the inner re-parse's `res` — it does *not* include
-                // the outer prefix the outer replay is about to splice
-                // in front. Without prepending that outer prefix, a
-                // later fail of *this* branch reconstructs its line
-                // from an empty prefix, dropping the outer captures
-                // entirely (the `[foo]:` LRD opener vanished from
-                // `syntax_test_markdown.md`'s `[foo]: /url` cases when
-                // a `link-def-attr-continuation` born inside the
-                // `link-def-title-continuation` replay later failed).
-                let prefix_ops = match &self.replay_prefix_ops {
-                    Some(outer) => {
-                        let mut combined = outer.clone();
-                        combined.extend(ops.iter().cloned());
-                        combined
-                    }
-                    None => ops.clone(),
-                };
-                let bp = BranchPoint {
-                    name: name.clone(),
-                    next_alternative: 1, // 0 is about to be pushed
-                    alternatives: alternatives.clone(),
-                    stack_snapshot: self.stack.clone(),
-                    proto_starts_snapshot: self.proto_starts.clone(),
-                    match_start: *start, // position before this match's advance
-                    trigger_match_start: match_start,
-                    pat_scope: pat.scope.clone(),
-                    line_number: bp_line_number,
-                    ops_snapshot_len: ops.len(),
-                    stack_depth: self.stack.len(),
-                    non_consuming_push_at_snapshot: *non_consuming_push_at,
-                    first_line_snapshot: self.first_line,
-                    with_prototype: pat.with_prototype.clone(),
-                    pending_lines_snapshot_len: bp_pending_lines_snapshot_len,
-                    escape_stack_snapshot: self.escape_stack.clone(),
-                    pop_count,
-                    prefix_ops,
-                    capture_ops: pat
-                        .captures
-                        .as_ref()
-                        .map(|m| build_capture_ops(m, &reg_match.regions))
-                        .unwrap_or_default(),
-                };
-                self.branch_points.push(bp);
-                if let Some(tracker) = self.inner_replay_max_depth.as_mut() {
-                    let last = self.branch_points.last().unwrap();
-                    if last.stack_depth > tracker.depth {
-                        tracker.depth = last.stack_depth;
-                        tracker.bp = Some(BpInfo {
-                            name: last.name.clone(),
-                            stack_depth: last.stack_depth,
-                            line_number: last.line_number,
-                            inner_producer: None,
-                        });
-                    }
-                }
-                // `pop: N + branch:` is **lookahead** per ST: the trigger
-                // token must NOT inherit the popped frames' meta_scope.
-                // Route through `Push { pop_count }` so the existing
-                // Push-with-pop lookahead path in `push_meta_ops` /
-                // `perform_op` handles both the deeper-pop emission and
-                // the runtime stack mutation.
-                synthetic_op = MatchOperation::Push {
-                    ctx_refs: vec![alternatives[0].clone()],
-                    pop_count,
-                };
-            } else {
-                unreachable!()
-            }
-        } else {
-            synthetic_op = pat.operation.clone();
-        }
-
-        let op_to_use = if is_branch {
-            &synthetic_op
-        } else {
-            &pat.operation
-        };
-
-        self.push_meta_ops(true, match_start, level_context, op_to_use, syntax_set, ops)?;
-        for s in &pat.scope {
-            ops.push((match_start, ScopeStackOp::Push(*s)));
-        }
-        let capture_ops = pat
-            .captures
-            .as_ref()
-            .map(|m| build_capture_ops(m, &reg_match.regions))
-            .unwrap_or_default();
-        ops.extend(capture_ops.iter().cloned());
-        if !pat.scope.is_empty() {
-            ops.push((match_end, ScopeStackOp::Pop(pat.scope.len())));
-        }
-        self.push_meta_ops(false, match_end, level_context, op_to_use, syntax_set, ops)?;
-
-        if is_branch {
-            // Execute the synthetic Push through perform_op
-            let synthetic_pat = MatchPattern::new(
-                pat.has_captures,
-                pat.regex.regex_str().to_string(),
-                pat.scope.clone(),
-                pat.captures.clone(),
-                synthetic_op,
-                pat.with_prototype.clone(),
-            );
-            self.perform_op(line, &reg_match.regions, &synthetic_pat, syntax_set)
-        } else {
-            self.perform_op(line, &reg_match.regions, pat, syntax_set)
-        }
-    }
-
-    /// Merge a cross-line replay's per-line corrected ops into `flushed_ops`.
-    ///
-    /// Multiple cross-line fails can fire on a single `parse_line` call (e.g.
-    /// Java line 624 with two live branches from line 615, both snapshotted at
-    /// `pending_lines_snapshot_len = 0`). Each fail's replay covers
-    /// `pending_lines[snap..pending_lines.len())`. A naive `extend` leaves
-    /// `flushed_ops` with duplicates — consumers of `ParseLineOutput::replayed`
-    /// index `replayed[i] ↔ pending_lines[i]` by `buf_len - replayed.len()`,
-    /// so duplicates misalign every byte offset.
-    ///
-    /// Composition rule, given current start `a` and new fail's `snap`:
-    /// - `snap <= a`: new fail supersedes everything; replace.
-    /// - `snap > a`: keep `[a..snap)` from prior fails, replace `[snap..N)`.
-    ///
-    /// `new_per_slot_bp` records the BP attribution for each new slot;
-    /// stored per slot in `flushed_ops_bp_per_slot` so subsequent merges
-    /// can compare against the BP that owns each individual slot rather
-    /// than a single BP for the whole buffer. When two cross-line fails
-    /// on the same `parse_line` both contribute slots, each slot's BP
-    /// reflects which fail wrote it. When `prefer_inner_replay_corrections`
-    /// substituted a slot from inner replay, that slot's `BpInfo`
-    /// surfaces the inner producer via `inner_producer` — finer-grained
-    /// attribution than rolling everything up to the outer BP.
-    fn merge_flushed(
-        &mut self,
-        snap: usize,
-        new_ops: Vec<Vec<(usize, ScopeStackOp)>>,
-        new_per_slot_bp: Vec<BpInfo>,
-    ) {
-        let new_len = new_ops.len();
-        debug_assert_eq!(new_len, new_per_slot_bp.len());
-        // For SnapLeStart's per-slot line-number discriminator and
-        // for trace events, we need the new BP's line_number — outer
-        // attribution is uniform across the call (slots may carry
-        // distinct inner_producers but share the outer's line/depth),
-        // so reading slot 0 is representative. Fall back to `0` for
-        // an empty merge (no slots to compare against anyway).
-        let new_bp_line_number = new_per_slot_bp.first().map(|b| b.line_number).unwrap_or(0);
-        match self.flushed_ops_start {
-            None => {
-                self.flushed_ops = new_ops;
-                self.flushed_ops_start = Some(snap);
-                self.flushed_ops_bp_per_slot = new_per_slot_bp;
-            }
-            Some(start) if snap <= start => {
-                // Per-slot line-number discriminator: a prior slot
-                // established by a BP from a later line is finer-grained
-                // than the wrapping earlier-line retry that's now
-                // overwriting. Preserve such slots; let equal/earlier-line
-                // priors fall through to the overwrite (peer-BP arbitration
-                // at the same line, or a fresher finding).
-                let prior_len = self.flushed_ops.len();
-                let overlap_end = (start + prior_len).min(snap + new_len);
-                let mut composed_ops = new_ops;
-                let mut composed_bp = new_per_slot_bp;
-                for abs in start..overlap_end {
-                    let prior_idx = abs - start;
-                    let new_idx = abs - snap;
-                    if self.flushed_ops_bp_per_slot[prior_idx].line_number > new_bp_line_number {
-                        composed_ops[new_idx] = std::mem::take(&mut self.flushed_ops[prior_idx]);
-                        composed_bp[new_idx] = self.flushed_ops_bp_per_slot[prior_idx].clone();
-                    }
-                }
-                self.flushed_ops = composed_ops;
-                self.flushed_ops_start = Some(snap);
-                self.flushed_ops_bp_per_slot = composed_bp;
-            }
-            Some(start) => {
-                // Per-slot effective-depth discriminator in the
-                // [snap..overlap_end) range. A prior slot's effective
-                // producer depth is the max of its own `stack_depth`
-                // and any recursive `inner_producer` chain; preserve
-                // the prior when its effective depth strictly exceeds
-                // the new BP's. Cluster A's killing inner-most
-                // `groups(d=4,l=3)` overwrite of `AQI(d=5,l=1)` slots
-                // (eff prior=5 > new=4) gets preserved here; rolled-up
-                // later merges keep firing because the surviving
-                // slots' `inner_producer` chains carry the AQI depth
-                // forward
-                // (`cross_line_chained_fail_pushes_target_meta_scope_on_continuation_line`,
-                // probe prose at `parser.rs:10270`+).
-                let prior_ops = std::mem::take(&mut self.flushed_ops);
-                let prior_bp = std::mem::take(&mut self.flushed_ops_bp_per_slot);
-                let prior_len = prior_ops.len();
-                let keep = snap - start;
-                let overlap_end = (start + prior_len).min(snap + new_len);
-                let overlap_count = overlap_end.saturating_sub(snap);
-                let mut composed_ops: Vec<Vec<(usize, ScopeStackOp)>> =
-                    Vec::with_capacity(keep + new_len);
-                let mut composed_bp: Vec<BpInfo> = Vec::with_capacity(keep + new_len);
-                let mut prior_ops_iter = prior_ops.into_iter();
-                let mut prior_bp_iter = prior_bp.into_iter();
-                let mut new_ops_iter = new_ops.into_iter();
-                let mut new_bp_iter = new_per_slot_bp.into_iter();
-                for _ in 0..keep {
-                    composed_ops.push(prior_ops_iter.next().unwrap());
-                    composed_bp.push(prior_bp_iter.next().unwrap());
-                }
-                for _ in 0..overlap_count {
-                    let new_op = new_ops_iter.next().unwrap();
-                    let new_bp = new_bp_iter.next().unwrap();
-                    let prior_op = prior_ops_iter.next().unwrap();
-                    let prior_bp_slot = prior_bp_iter.next().unwrap();
-                    if Self::effective_producer_depth(&prior_bp_slot)
-                        > Self::effective_producer_depth(&new_bp)
-                    {
-                        composed_ops.push(prior_op);
-                        composed_bp.push(prior_bp_slot);
-                    } else {
-                        composed_ops.push(new_op);
-                        composed_bp.push(new_bp);
-                    }
-                }
-                composed_ops.extend(new_ops_iter);
-                composed_bp.extend(new_bp_iter);
-                self.flushed_ops = composed_ops;
-                // SnapGtStart preserves the prior buffer's start
-                // index — `keep` prior slots remain before the
-                // overlap/new tail. Original truncate-and-extend
-                // didn't touch `flushed_ops_start` either.
-                self.flushed_ops_start = Some(start);
-                self.flushed_ops_bp_per_slot = composed_bp;
-            }
-        }
-    }
-
-    fn effective_producer_depth(bp: &BpInfo) -> usize {
-        let mut d = bp.stack_depth;
-        let mut cur = bp.inner_producer.as_deref();
-        while let Some(c) = cur {
-            if c.stack_depth > d {
-                d = c.stack_depth;
-            }
-            cur = c.inner_producer.as_deref();
-        }
-        d
-    }
-
-    /// Replace `replayed_ops[i]` with `inner_ops` for overlapping
-    /// indices. Only fires when the inner BP was created at a stack
-    /// depth less-than-or-equal to the outer BP's depth — i.e. the
-    /// inner BP is a sibling-or-shallower correction at the same
-    /// structural level, not a nested BP firing inside outer's resolved
-    /// alternative. The "deeper inner" case is the multigen16
-    /// regression seat: outer = `class-members` at depth 4, inner =
-    /// `object-type` at depth 9 (nested inside outer's resolved field
-    /// alt). Preferring its corrections doubled `meta.field.type.java`
-    /// on `Java/syntax_test_java.java:3462+` because outer's full-line
-    /// ops correctly emit one `meta.field.type` and inner's reparse
-    /// adds another.
-    ///
-    /// The depth-equal case is the original `@A.B\n(par=1)\nenum E {}`
-    /// regression seat from PR #663: outer and inner are both
-    /// `declarations` at depth 3 — sibling resolutions of the same
-    /// branch family, and inner's `name`-alt CORRECTLY supersedes
-    /// outer's locally-computed `path`-alt freeze.
-    fn prefer_inner_replay_corrections(
-        &mut self,
-        outer_snap: usize,
-        replayed_ops: &mut [Vec<(usize, ScopeStackOp)>],
-        replayed_per_slot_bp: &mut [BpInfo],
-        inner_ops: &[Vec<(usize, ScopeStackOp)>],
-        inner_start: usize,
-        outer_bp: &BpInfo,
-        inner_bp: Option<&BpInfo>,
-        inner_corrections_bp: &[BpInfo],
-    ) {
-        // Two regimes:
-        //
-        // 1. Inner at outer's exact depth (sibling resolution of the
-        //    same branch family) or exactly one deeper (immediate
-        //    refinement, e.g. outer=declarations(3),
-        //    inner=annotation-identifier(4) on the same line — inner
-        //    brings `meta.path.java` from the qualified-identifier alt
-        //    that outer's locally-computed parse drops). Always
-        //    substitute.
-        //
-        // 2. Inner more than one deeper (multigen16-style: outer
-        //    class-members(4), inner object-type(11)). Inner is
-        //    nested INSIDE outer's resolved alt, and its reparse may
-        //    push atoms outer's alt already provides — substituting
-        //    blindly would double them
-        //    (`deeper_inner_bp_correction_does_not_double_outer_meta_scope`).
-        //    Substitute only when the inner's ops are an `immediately-
-        //    pop`-style tail-extension of outer's: identical prefix +
-        //    one or more `Pop` ops appended at positions outer
-        //    already covers. That's what an `immediately-pop`-alt
-        //    failover commits at the inner BP's trigger position
-        //    when outer's per-line replay terminated before the
-        //    nested fail fired (cluster-C
-        //    `multi_line_annotation_pops_meta_scope_at_eol`).
-        //
-        // Inner shallower than outer is also skipped: such inners
-        // are fresh top-level-style BPs whose commitment is
-        // structurally less specific than outer's and would overwrite
-        // outer's correct refined parse with a coarser one.
-        let depth_diff =
-            inner_bp.map(|inner| inner.stack_depth as isize - outer_bp.stack_depth as isize);
-        let in_depth_window = matches!(depth_diff, Some(0) | Some(1));
-        let allow_deep_extension = matches!(depth_diff, Some(d) if d > 1);
-        debug_assert_eq!(replayed_ops.len(), replayed_per_slot_bp.len());
-        // Helper closure: build the per-slot BpInfo recorded for a
-        // slot we just substituted from inner replay. Outer attribution
-        // is preserved (so subsequent line/depth comparisons against
-        // the outer wrapping BP keep working) and the inner producer
-        // is recorded under `inner_producer` for finer-grained
-        // discrimination at the next merge. `slot_inner_bp` already
-        // carries any deeper nested attribution recursively.
-        let attribute_substituted = |slot_inner_bp: Option<&BpInfo>| -> BpInfo {
-            let producer = slot_inner_bp
-                .cloned()
-                .or_else(|| inner_bp.cloned())
-                .expect("substitution path requires an inner producer");
-            BpInfo {
-                name: outer_bp.name.clone(),
-                stack_depth: outer_bp.stack_depth,
-                line_number: outer_bp.line_number,
-                inner_producer: Some(Box::new(producer)),
-            }
-        };
-        for (i, (outer_local, slot_bp_attr)) in replayed_ops
-            .iter_mut()
-            .zip(replayed_per_slot_bp.iter_mut())
-            .enumerate()
-        {
-            let global_i = outer_snap + i;
-            if global_i < inner_start {
-                continue;
-            }
-            let inner_idx = global_i - inner_start;
-            let slot_inner_bp = inner_corrections_bp.get(inner_idx);
-            // Per-slot line-number escape valve: an inner correction
-            // whose BP was created on a later line than outer's wrapping
-            // BP carries finer-grained context that outer's recompute
-            // can't reproduce. Substitute regardless of depth.
-            let slot_line_overrides = matches!(
-                slot_inner_bp,
-                Some(sib) if sib.line_number > outer_bp.line_number
-            );
-            if let Some(corrected) = inner_ops.get(inner_idx) {
-                if in_depth_window {
-                    *outer_local = corrected.clone();
-                    *slot_bp_attr = attribute_substituted(slot_inner_bp);
-                } else if slot_line_overrides {
-                    *outer_local = corrected.clone();
-                    *slot_bp_attr = attribute_substituted(slot_inner_bp);
-                } else if allow_deep_extension && Self::inner_extends_outer(outer_local, corrected)
-                {
-                    *outer_local = corrected.clone();
-                    *slot_bp_attr = attribute_substituted(slot_inner_bp);
-                } else {
-                    // Iter 7 substitution path: when SkippedDeepNonExtension
-                    // would fire and `is_replace_shape` matches, substitute
-                    // outer_local with inner's corrected ops. If the
-                    // substitution would over-push a `meta.*` atom not
-                    // already on the running shadow stack (G2 discriminator
-                    // — see deferral prose at `parser.rs:11363+`), append a
-                    // compensating Pop(1) immediately after each over-pushed
-                    // Push so the consumer's stack mirror sees the
-                    // intended single push without doubling.
-                    let (_, outer_first, inner_first) =
-                        Self::ops_divergence(outer_local, corrected);
-                    let outer_meta = Self::collect_meta_scopes(outer_local);
-                    let inner_meta = Self::collect_meta_scopes(corrected);
-                    let would_be_skipped_deep =
-                        inner_bp.is_some() && !matches!(depth_diff, Some(d) if d < 0);
-                    let do_substitute = would_be_skipped_deep
-                        && Self::is_replace_shape(
-                            &outer_first,
-                            &inner_first,
-                            &outer_meta,
-                            &inner_meta,
-                        );
-                    if do_substitute {
-                        let outer_pre_substitution = outer_local.clone();
-                        *outer_local = corrected.clone();
-                        *slot_bp_attr = attribute_substituted(slot_inner_bp);
-                        // Per-meta atom over-push: positive entries of
-                        // (inner_net - outer_net). When non-empty, inner
-                        // pushes a meta atom that outer doesn't.
-                        let outer_net = Self::net_meta_delta(&outer_pre_substitution);
-                        let inner_net = Self::net_meta_delta(corrected);
-                        let mut over_push: HashMap<String, usize> = HashMap::new();
-                        for (atom, inner_c) in &inner_net {
-                            let outer_c = outer_net.get(atom).copied().unwrap_or(0);
-                            if *inner_c > outer_c {
-                                over_push.insert(atom.clone(), inner_c - outer_c);
-                            }
-                        }
-                        // G2 gate: skip comp-pop when any over-push atom
-                        // is already on the running shadow stack — that
-                        // signals inner is legitimately preserving an
-                        // existing meta scope (not introducing a fresh
-                        // one), so popping it would break the consumer.
-                        if !over_push.is_empty() {
-                            let would_double = self.shadow.as_slice().iter().any(|s| {
-                                let n = s.build_string();
-                                n.starts_with("meta.") && over_push.contains_key(&n)
-                            });
-                            if !would_double {
-                                let total_extra: usize = over_push.values().sum();
-                                let mut remaining = over_push;
-                                let mut filtered: Vec<(usize, ScopeStackOp)> =
-                                    Vec::with_capacity(outer_local.len() + total_extra);
-                                for entry in outer_local.iter() {
-                                    filtered.push(entry.clone());
-                                    if let (pos, ScopeStackOp::Push(s)) = (entry.0, &entry.1) {
-                                        let name = s.build_string();
-                                        if let Some(rem) = remaining.get_mut(&name) {
-                                            if *rem > 0 {
-                                                filtered.push((pos, ScopeStackOp::Pop(1)));
-                                                *rem -= 1;
-                                            }
-                                        }
-                                    }
-                                }
-                                *outer_local = filtered;
-                            }
-                        }
-                        // Iter 9 deferred sub-B's outer-side comp-pop:
-                        // iter-8 analysis flagged outer over-pushes
-                        // (`meta.function.*` family inner correctly
-                        // omits) but the symmetric predicate widening
-                        // surfaced inner-BP stack-base misalignment
-                        // first — see iter-9 prose at `parser.rs:11968+`
-                        // for the deferral rationale and iter 9.5
-                        // discriminator requirements.
-                    } else {
-                    }
-                }
-            } else {
-            }
-        }
-    }
-
-    /// Length of the longest equal prefix between `outer` and `inner`,
-    /// plus a summary of each side's first op past that prefix (or
-    /// `None` if that side is exhausted at the prefix boundary). Used
-    /// by Iter 7's `is_replace_shape` substitution predicate (production)
-    /// and the `SubstitutionShapeAnalysis` trace event (debug).
-    fn ops_divergence(
-        outer: &[(usize, ScopeStackOp)],
-        inner: &[(usize, ScopeStackOp)],
-    ) -> (usize, Option<OpSummary>, Option<OpSummary>) {
-        let mut prefix = 0usize;
-        let max = outer.len().min(inner.len());
-        while prefix < max && outer[prefix] == inner[prefix] {
-            prefix += 1;
-        }
-        let outer_div = outer.get(prefix).map(Self::summarize_op);
-        let inner_div = inner.get(prefix).map(Self::summarize_op);
-        (prefix, outer_div, inner_div)
-    }
-
-    /// Collect the names of every `Push(scope)` whose scope name
-    /// begins with `meta.`, in occurrence order. These are the
-    /// meta-scope frames the ops would push onto the actual scope
-    /// stack when applied. Used by Iter 7's substitution predicate
-    /// (`is_replace_shape`) and the `SubstitutionShapeAnalysis` trace.
-    fn collect_meta_scopes(ops: &[(usize, ScopeStackOp)]) -> Vec<String> {
-        let mut out = Vec::new();
-        for (_, op) in ops {
-            if let ScopeStackOp::Push(scope) = op {
-                let s = scope.build_string();
-                if s.starts_with("meta.") {
-                    out.push(s);
-                }
-            }
-        }
-        out
-    }
-
-    fn summarize_op(entry: &(usize, ScopeStackOp)) -> OpSummary {
-        let (pos, op) = entry;
-        let (kind, scope) = match op {
-            ScopeStackOp::Push(s) => ("push", s.build_string()),
-            ScopeStackOp::Pop(_) => ("pop", String::new()),
-            ScopeStackOp::Clear(_) => ("clear", String::new()),
-            ScopeStackOp::Restore => ("restore", String::new()),
-            ScopeStackOp::Noop => ("noop", String::new()),
-        };
-        OpSummary {
-            pos: *pos,
-            kind,
-            scope,
-        }
-    }
-
-    /// True when both outer and inner have at least one diverging op past their common prefix
-    /// and `inner_meta_scopes` contains a `meta.*` atom not in `outer_meta_scopes`. Used to
-    /// gate the substitution path in `prefer_inner_replay_corrections`.
-    fn is_replace_shape(
-        outer_div: &Option<OpSummary>,
-        inner_div: &Option<OpSummary>,
-        outer_meta_scopes: &[String],
-        inner_meta_scopes: &[String],
-    ) -> bool {
-        if outer_div.is_none() || inner_div.is_none() {
-            return false;
-        }
-        inner_meta_scopes
-            .iter()
-            .any(|s| s.starts_with("meta.") && !outer_meta_scopes.contains(s))
-    }
-
-    /// Simulate `ops` on a fresh scope-name stack and return the multiset of `meta.*` atoms
-    /// remaining at the end. `Push`/`Pop(N)` are simulated literally; other op variants are
-    /// treated as no-ops. Used to compute the per-atom net push delta between outer and inner
-    /// ops at the comp-pop site in `prefer_inner_replay_corrections`.
-    fn net_meta_delta(ops: &[(usize, ScopeStackOp)]) -> HashMap<String, usize> {
-        let mut sim_stack: Vec<String> = Vec::new();
-        for (_, op) in ops {
-            match op {
-                ScopeStackOp::Push(s) => sim_stack.push(s.build_string()),
-                ScopeStackOp::Pop(n) => {
-                    for _ in 0..*n {
-                        sim_stack.pop();
-                    }
-                }
-                _ => {}
-            }
-        }
-        let mut counts: HashMap<String, usize> = HashMap::new();
-        for s in &sim_stack {
-            if s.starts_with("meta.") {
-                *counts.entry(s.clone()).or_insert(0) += 1;
-            }
-        }
-        counts
-    }
-
-    /// True when `inner` starts with the entire `outer` op sequence and
-    /// the trailing extension contains only `Pop` ops at positions
-    /// `outer` already covers (i.e. the inner reparse adds end-of-line
-    /// pops that the outer's locally-computed parse did not emit, and
-    /// nothing else).
-    fn inner_extends_outer(
-        outer: &[(usize, ScopeStackOp)],
-        inner: &[(usize, ScopeStackOp)],
-    ) -> bool {
-        if inner.len() <= outer.len() {
-            return false;
-        }
-        if inner[..outer.len()] != *outer {
-            return false;
-        }
-        let max_pos = outer.last().map(|(p, _)| *p).unwrap_or(0);
-        inner[outer.len()..]
-            .iter()
-            .all(|(pos, op)| *pos <= max_pos && matches!(op, ScopeStackOp::Pop(_)))
-    }
-
-    /// Iter-19 (parser.rs:~12848 prose) discriminator: should cross-line
-    /// `handle_fail` rewind to BP-creation snapshot and force alt 5 on
-    /// re-entry? Composite signal — `class-members` name gate, not-
-    /// already-alt-5 loop guard, empty inner corrections, max-depth BP
-    /// strictly deeper than outer, and the first replayed slot's ops
-    /// contain a Push of `meta.function.return-type.java`. Iter-18's
-    /// per-(`bp_line`, alt) matrix witness: TRUE at row 5 (`@ 3394`
-    /// BENEFICIAL) + 2 safe-neutrals (2976, 2982); FALSE at all 12
-    /// harmful method-decl sites at all alts in `syntax_test_java.java`.
-    fn class_members_alt5_should_rewind(
-        outer_bp_info: &BpInfo,
-        next_alt_index: usize,
-        inner_corrections_empty: bool,
-        inner_replay_max: &MaxDepthSeen,
-        replayed_ops: &[Vec<(usize, ScopeStackOp)>],
-        return_type_scope: Scope,
-    ) -> bool {
-        if outer_bp_info.name != "class-members" {
-            return false;
-        }
-        if next_alt_index >= 5 {
-            return false;
-        }
-        if !inner_corrections_empty {
-            return false;
-        }
-        let max_bp = match inner_replay_max.bp.as_ref() {
-            Some(bp) => bp,
-            None => return false,
-        };
-        if max_bp.stack_depth <= outer_bp_info.stack_depth {
-            return false;
-        }
-        let first_slot = match replayed_ops.first() {
-            Some(slot) => slot,
-            None => return false,
-        };
-        first_slot.iter().any(|(_, op)| match op {
-            ScopeStackOp::Push(s) => return_type_scope.is_prefix_of(*s),
-            _ => false,
-        })
-    }
-
-    /// Handle a `fail` operation by rewinding to the named branch point.
-    /// Returns Ok(true) if backtracking happened (caller should continue from rewound position).
-    /// Returns Ok(false) if the fail had no effect.
-    fn handle_fail(
-        &mut self,
-        name: &str,
-        line: &str,
-        start: &mut usize,
-        non_consuming_push_at: &mut (usize, usize, usize),
-        ops: &mut Vec<(usize, ScopeStackOp)>,
-        search_cache: &mut SearchCache,
-        syntax_set: &SyntaxSet,
-    ) -> Result<bool, ParsingError> {
-        // Find the branch point by name (most recent first), skipping
-        // records whose alternative's pushed frame is no longer on
-        // the stack. The alternative lives at
-        // `bp.stack_depth - bp.pop_count + 1`, so `stack.len() >
-        // bp.stack_depth - bp.pop_count` means the frame is still
-        // present. Without this skip, a nested `branch_point` with
-        // the same name whose inner alternative popped cleanly would
-        // shadow an enclosing record via `rposition`, rewinding to
-        // the inner branch position instead of the outer one
-        // (Haskell's raw-string QQ `[r|[a-zA-Z]|]`).
-        let stack_len = self.stack.len();
-        let bp_index = self.branch_points.iter().rposition(|bp| {
-            bp.name == name && stack_len > bp.stack_depth.saturating_sub(bp.pop_count)
-        });
-        let bp_index = match bp_index {
-            Some(i) => i,
-            None => return Ok(false), // No such branch point, fail is no-op
-        };
-
-        // During a replay recursion, `cur_line` is the virtual replay
-        // line — without this override a same-line fail fired inside
-        // the re-parse would be misclassified as cross-line, and a
-        // fail on the outer line for a branch created during replay
-        // would be misclassified as same-line.
-        let cur_line = match &self.replay_ctx {
-            Some(ctx) => ctx.line_number,
-            None => self.line_number.saturating_sub(1),
-        };
-        let bp = &self.branch_points[bp_index];
-
-        // Check validity: not >128 lines old
-        if cur_line.saturating_sub(bp.line_number) > 128 {
-            let bp = self.branch_points.remove(bp_index);
-            self.warnings.push(format!(
-                "branch point '{}' expired (exceeded 128-line rewind limit)",
-                bp.name
-            ));
-            return Ok(false);
-        }
-
-        // Check validity: the alt frame is still on the stack. Mirrors the
-        // bp lookup predicate above (`stack_len > bp.stack_depth.saturating_sub(bp.pop_count)`)
-        // so a `pop: N + branch_point` whose snapshot captures the
-        // pre-pop depth doesn't false-positive here. Without subtracting
-        // `pop_count`, Java's `pop: 2 + branch_point: annotation-qualified-parameters`
-        // failed lookup with `self.stack.len() < bp.stack_depth` and the
-        // fail became a no-op, leaking `meta.annotation.identifier.java`
-        // into every nested-annotation extends path.
-        if self.stack.len() <= bp.stack_depth.saturating_sub(bp.pop_count) {
-            self.branch_points.remove(bp_index);
-            return Ok(false);
-        }
-
-        // Check if there are more alternatives
-        if bp.next_alternative >= bp.alternatives.len() {
-            // All alternatives exhausted: restore parser state to the
-            // pre-branch snapshot so the stuck alternative's pushed
-            // contexts and emitted ops are discarded, then advance
-            // past the branch_point match position by one character so
-            // we don't immediately re-enter the same branch_point and
-            // loop. Before this, the branch_point was silently removed
-            // while its last alternative's contexts remained on the
-            // stack — the cause of the "scope stack stays in
-            // `meta.interpolation.brace.shell`" cascade in Zsh when
-            // both `brace-interpolation-sequence` and
-            // `brace-interpolation-series` failed and there was no
-            // fallback alternative (Zsh explicitly excludes
-            // `brace-interpolation-fallback`).
-            //
-            // Cross-line exhaustion takes the same shape, plus a replay
-            // of the buffered lines under the pre-branch state so
-            // callers see corrected ops for lines they've already been
-            // handed. Without this, the unterminated TypeScript type
-            // expression at `sublimehq/Packages#3598`
-            // (`type x = { bar: (cb: (\n};`) left the inner
-            // `ts-type-function-parameter-list-body` on the stack
-            // forever, contaminating every subsequent line's scope
-            // stack with `meta.type.js, meta.group.js` — 274 cascading
-            // assertion failures in `syntax_test_typescript.ts`.
-            let is_cross_line = bp.line_number < cur_line;
-            let bp_line_number = bp.line_number;
-            let stack_snapshot = bp.stack_snapshot.clone();
-            let proto_starts_snapshot = bp.proto_starts_snapshot.clone();
-            let escape_stack_snapshot = bp.escape_stack_snapshot.clone();
-            let first_line_snapshot = bp.first_line_snapshot;
-            let non_consuming_push_at_snapshot = bp.non_consuming_push_at_snapshot;
-            let ops_snapshot_len = bp.ops_snapshot_len;
-            let match_start_pos = bp.match_start;
-            let pending_lines_snapshot_len = bp.pending_lines_snapshot_len;
-            let prefix_ops = bp.prefix_ops.clone();
-            let outer_bp_info = BpInfo {
-                name: bp.name.clone(),
-                stack_depth: bp.stack_depth,
-                line_number: bp.line_number,
-                inner_producer: None,
-            };
-            self.branch_points.remove(bp_index);
-
-            self.stack = stack_snapshot;
-            self.proto_starts = proto_starts_snapshot;
-            self.escape_stack = escape_stack_snapshot;
-            self.first_line = first_line_snapshot;
-            *non_consuming_push_at = non_consuming_push_at_snapshot;
-            ops.truncate(ops_snapshot_len.min(ops.len()));
-
-            if is_cross_line {
-                // Re-parse each buffered line under the restored (pre-branch)
-                // state so `parse_line` can surface the corrected ops via
-                // `ParseLineOutput::replayed`. The first buffered line is the
-                // branch-creation line: emit its saved `prefix_ops` (the ops
-                // emitted before the branch match) verbatim, then advance past
-                // the branch match by one character before resuming — otherwise
-                // the same branch_point would fire again at the original match
-                // position and we'd loop.
-                //
-                // Keep `pending_lines` intact (don't drain): if an outer
-                // branch_point on this same line also fails after this
-                // exhaustion replay, its own replay needs access to the same
-                // buffered lines.
-                let truncated_lines: Vec<String> =
-                    self.pending_lines[pending_lines_snapshot_len..].to_vec();
-                // Save prior flushed_ops state and clear it so any nested
-                // cross-line fails firing during the replay loop below
-                // write into a clean slot we can detect afterward.
-                let saved_flushed = std::mem::take(&mut self.flushed_ops);
-                let saved_flushed_start = self.flushed_ops_start.take();
-                let saved_flushed_bp = std::mem::take(&mut self.flushed_ops_bp_per_slot);
-                // Install a fresh max-depth tracker for the duration of
-                // this inner replay; restore the outer tracker (if any)
-                // when we're done. Mirrors `saved_flushed`'s nesting.
-                let saved_max_depth = self.inner_replay_max_depth.replace(MaxDepthSeen::default());
-                let mut replayed_ops: Vec<Vec<(usize, ScopeStackOp)>> =
-                    Vec::with_capacity(truncated_lines.len());
-                for (i, replay_line) in truncated_lines.iter().enumerate() {
-                    // Tag branches created during this iteration with
-                    // the replay line's identity, not the outer line's.
-                    let prev_replay_ctx = self.replay_ctx.replace(ReplayCtx {
-                        line_number: bp_line_number + i,
-                        pending_lines_snapshot_offset: pending_lines_snapshot_len + i,
-                    });
-                    // No new-alt construction here (all alternatives
-                    // exhausted), so the first-line prefix is just
-                    // `prefix_ops`. Surface it to inner branch creations
-                    // so their `prefix_ops` keeps the outer captures.
-                    let prev_replay_prefix = if i == 0 {
-                        self.replay_prefix_ops.replace(prefix_ops.clone())
-                    } else {
-                        self.replay_prefix_ops.take()
-                    };
-                    let inner_result = if i == 0 {
-                        self.skipped_branches
-                            .push((match_start_pos, name.to_string()));
-                        self.parse_line_inner_from(replay_line, syntax_set, match_start_pos)
-                    } else {
-                        self.parse_line_inner(replay_line, syntax_set)
-                    };
-                    self.replay_ctx = prev_replay_ctx;
-                    self.replay_prefix_ops = prev_replay_prefix;
-                    let tail_ops = inner_result?;
-                    let line_ops = if i == 0 {
-                        let mut first_line_ops = prefix_ops.clone();
-                        first_line_ops.extend(tail_ops);
-                        first_line_ops
-                    } else {
-                        tail_ops
-                    };
-                    replayed_ops.push(line_ops);
-                }
-                // Capture inner corrections (if any), restore prior state,
-                // then prefer the inner corrections for overlapping indices.
-                let inner_corrections = std::mem::take(&mut self.flushed_ops);
-                let inner_corrections_start = self.flushed_ops_start.take();
-                let inner_corrections_bp = std::mem::take(&mut self.flushed_ops_bp_per_slot);
-                self.flushed_ops = saved_flushed;
-                self.flushed_ops_start = saved_flushed_start;
-                self.flushed_ops_bp_per_slot = saved_flushed_bp;
-                let _inner_replay_max =
-                    std::mem::take(&mut self.inner_replay_max_depth).unwrap_or_default();
-                self.inner_replay_max_depth = saved_max_depth;
-                let mut replayed_per_slot_bp: Vec<BpInfo> =
-                    vec![outer_bp_info.clone(); replayed_ops.len()];
-                if let Some(start) = inner_corrections_start {
-                    if !inner_corrections.is_empty() {
-                        let inner_bp_first = inner_corrections_bp.first().cloned();
-                        self.prefer_inner_replay_corrections(
-                            pending_lines_snapshot_len,
-                            &mut replayed_ops,
-                            &mut replayed_per_slot_bp,
-                            &inner_corrections,
-                            start,
-                            &outer_bp_info,
-                            inner_bp_first.as_ref(),
-                            &inner_corrections_bp,
-                        );
-                    }
-                }
-                self.merge_flushed(
-                    pending_lines_snapshot_len,
-                    replayed_ops,
-                    replayed_per_slot_bp,
-                );
-
-                // Restart the current line from the beginning under the
-                // restored state.
-                ops.clear();
-                *start = 0;
-                *non_consuming_push_at = (0, 0, 0);
-                search_cache.clear();
-                return Ok(true);
-            }
-
-            // Same-line exhaustion: rewind the cursor to the BP's
-            // original position and record the branch_point's name so
-            // subsequent `find_best_match` calls at that position skip
-            // the same-name Branch pattern. This lets the parent
-            // context's NEXT rule fire instead of advancing past the
-            // lookahead match — mirrors ST's branch-point exhaustion
-            // semantics. The previous behaviour (advance one char) let
-            // stale keyword rules match in the middle of identifiers,
-            // e.g. `package` inside `$package` after `declarations`
-            // exhausted on the leading `$`. If `find_best_match`
-            // returns nothing at this cursor, `parse_next_token`'s
-            // no-match fallback advances one char as a last resort.
-            self.skipped_branches
-                .push((match_start_pos, name.to_string()));
-            *start = match_start_pos;
-            search_cache.clear();
-            return Ok(true);
-        }
-
-        // Determine if this is a cross-line fail (branch was created on a previous line).
-        let is_cross_line = bp.line_number < cur_line;
-
-        // Extract everything we need from bp before mutating self.
-        let bp_line_number = bp.line_number;
-        let next_alt_index = bp.next_alternative;
-        let next_alt = bp.alternatives[next_alt_index].clone();
-        let match_start_pos = bp.match_start;
-        let trigger_match_start = bp.trigger_match_start;
-        let trigger_pat_scope = bp.pat_scope.clone();
-        let trigger_capture_ops = bp.capture_ops.clone();
-        let stack_snapshot = bp.stack_snapshot.clone();
-        let proto_starts_snapshot = bp.proto_starts_snapshot.clone();
-        let first_line_snapshot = bp.first_line_snapshot;
-        let non_consuming_push_at_snapshot = bp.non_consuming_push_at_snapshot;
-        let ops_snapshot_len = bp.ops_snapshot_len;
-        let pending_lines_snapshot_len = bp.pending_lines_snapshot_len;
-        let escape_stack_snapshot = bp.escape_stack_snapshot.clone();
-        let prefix_ops = bp.prefix_ops.clone();
-        let outer_bp_info = BpInfo {
-            name: bp.name.clone(),
-            stack_depth: bp.stack_depth,
-            line_number: bp.line_number,
-            inner_producer: None,
-        };
-        // bp borrow ends here.
-
-        let pop_count = self.branch_points[bp_index].pop_count;
-
-        // Restore parser state to the snapshot.
-        // Keep `stack_snapshot` available — the same-line fix below needs
-        // it to compute popped-context meta_scope clearance via
-        // `push_meta_ops`.
-        self.stack = stack_snapshot.clone();
-        self.proto_starts = proto_starts_snapshot.clone();
-        self.escape_stack = escape_stack_snapshot.clone();
-        self.first_line = first_line_snapshot;
-        *non_consuming_push_at = non_consuming_push_at_snapshot;
-
-        // Update the branch point record before popping/pushing
-        // (must happen before the pop which may invalidate indices).
-        self.branch_points[bp_index].next_alternative = next_alt_index + 1;
-
-        // For pop + branch: re-pop the contexts (snapshot was taken pre-pop).
-        if pop_count > 0 {
-            for _ in 0..pop_count {
-                self.stack.pop();
-            }
-        }
-
-        // Push the next alternative onto the stack.
-        let with_prototype = self.branch_points[bp_index].with_prototype.clone();
-        let context_id = next_alt.id()?;
-        let captures = None; // no captures available at rewind time
-
-        let proto_ids = match with_prototype {
-            Some(ref p) => vec![p.id()?],
-            None => Vec::new(),
-        };
-
-        self.stack.push(StateLevel {
-            context: context_id,
-            prototypes: proto_ids,
-            captures,
-        });
-
-        if is_cross_line {
-            // Cross-line fail: the ops for lines since the branch was created
-            // have already been returned to callers.  Re-parse those lines under
-            // the new alternative and store the corrected ops in `flushed_ops`
-            // so that `parse_line` can surface them via `ParseLineOutput::replayed`.
-            //
-            // The first buffered line is the branch-creation line. Its
-            // pre-branch prefix (cols 0..trigger_match_start) was correctly
-            // parsed under the *pre-branch* state — not the new alternative.
-            // Re-parsing it from column 0 with the new alternative on the
-            // stack would misattribute that prefix to the new alternative's
-            // rules (observed on multi-line SQL `LIKE … ESCAPE …`: every
-            // non-whitespace before `LIKE` fires `else-pop` in the
-            // escape-alternative, derailing the stack). Instead, reuse the
-            // prefix_ops saved at branch-creation time, manually emit the
-            // branch trigger's pat.scope and the new alternative's meta
-            // scope ops, then resume parsing from match_end with the new
-            // alternative on the stack via `parse_line_inner_from`.
-            // Keep `pending_lines` intact (don't drain): if a second branch_point
-            // on the current line also fails after this retry, its own replay
-            // needs access to the same buffered lines. Nested branches from the
-            // same earlier line share the buffer.
-            let truncated_lines: Vec<String> =
-                self.pending_lines[pending_lines_snapshot_len..].to_vec();
-
-            // Compose the first replayed line's prefix (outer prefix_ops +
-            // new-alt meta/pat/capture/meta_content emission) up front so a
-            // branch_point born inside the inner re-parse can inherit it
-            // via `self.replay_prefix_ops`. Built once per fail; cloned
-            // and extended with `tail_ops` to form the final line_ops.
-            //
-            // Use `push_meta_ops` with a synthetic Set/Push for the
-            // new alternative — same path the same-line fail above and
-            // the original branch creation take — so both the new
-            // alternative's own meta scopes AND the popped contexts'
-            // meta_scope/mcs clearance Pop (for `pop: N + branch_point`,
-            // N > 0) get re-emitted. A bespoke re-emit of just
-            // `context.meta_scope` / `context.meta_content_scope` is
-            // missing the popped-contexts Pop, leaving Java's
-            // `pop: 2 + branch_point: annotation-qualified-parameters`
-            // crossing a line boundary with both the popped context's
-            // meta_scope (`meta.annotation.identifier.java`) AND the
-            // outer declaration's meta_scope (`meta.enum.java` /
-            // `meta.class.java` / `meta.interface.java`) leaked on the
-            // stack — cascading across 8000+ lines past the multi-line
-            // annotation-modified declaration at lines 2260-2297 of
-            // `syntax_test_java.java`.
-            //
-            // `push_meta_ops` reads `self.stack` to compute the popped
-            // contexts' scope atoms, so swap in `stack_snapshot` (pre-pop
-            // state captured at branch creation) for the duration of the
-            // calls — `self.stack` currently holds the post-set state
-            // (alt N already pushed).
-            let mut first_line_prefix = prefix_ops.clone();
-            let synthetic_op_alt_n = MatchOperation::Push {
-                ctx_refs: vec![next_alt.clone()],
-                pop_count,
-            };
-            let level_ctx_id = stack_snapshot.last().map(|l| l.context);
-            let post_set_stack = std::mem::replace(&mut self.stack, stack_snapshot.clone());
-            if let Some(level_ctx_id) = level_ctx_id {
-                let level_context = syntax_set.get_context(&level_ctx_id)?;
-                self.push_meta_ops(
-                    true,
-                    trigger_match_start,
-                    level_context,
-                    &synthetic_op_alt_n,
-                    syntax_set,
-                    &mut first_line_prefix,
-                )?;
-                for scope in &trigger_pat_scope {
-                    first_line_prefix.push((trigger_match_start, ScopeStackOp::Push(*scope)));
-                }
-                first_line_prefix.extend(trigger_capture_ops.iter().cloned());
-                if !trigger_pat_scope.is_empty() {
-                    first_line_prefix
-                        .push((match_start_pos, ScopeStackOp::Pop(trigger_pat_scope.len())));
-                }
-                self.push_meta_ops(
-                    false,
-                    match_start_pos,
-                    level_context,
-                    &synthetic_op_alt_n,
-                    syntax_set,
-                    &mut first_line_prefix,
-                )?;
-            }
-            self.stack = post_set_stack;
-
-            // Snapshot the branch_points so the post-replay
-            // `class_members_alt5_should_rewind` discriminator can restore them
-            // for the inline alt-5 re-run below. The captured `next_alternative`
-            // is overwritten to 5 inside the rewind block, so its value here
-            // doesn't matter.
-            let branch_points_snapshot_for_rewind: Vec<BranchPoint> = self.branch_points.clone();
-
-            // Save prior flushed_ops state and clear it so any nested
-            // cross-line fails firing during the replay loop below write
-            // into a clean slot we can detect afterward.
-            let saved_flushed = std::mem::take(&mut self.flushed_ops);
-            let saved_flushed_start = self.flushed_ops_start.take();
-            let saved_flushed_bp = std::mem::take(&mut self.flushed_ops_bp_per_slot);
-            // Install a fresh max-depth tracker for the duration of
-            // this inner replay; restore the outer tracker (if any)
-            // when we're done. Mirrors `saved_flushed`'s nesting.
-            let saved_max_depth = self.inner_replay_max_depth.replace(MaxDepthSeen::default());
-
-            let mut replayed_ops: Vec<Vec<(usize, ScopeStackOp)>> =
-                Vec::with_capacity(truncated_lines.len());
-            for (i, replay_line) in truncated_lines.iter().enumerate() {
-                // Tag branches created during this iteration with the
-                // replay line's identity, not the outer line's.
-                let prev_replay_ctx = self.replay_ctx.replace(ReplayCtx {
-                    line_number: bp_line_number + i,
-                    pending_lines_snapshot_offset: pending_lines_snapshot_len + i,
-                });
-                // Expose the first-line prefix so a branch_point born
-                // during this line's re-parse anchors its `prefix_ops`
-                // to the full line state — outer captures included.
-                // Subsequent replayed lines start fresh.
-                let prev_replay_prefix = if i == 0 {
-                    self.replay_prefix_ops.replace(first_line_prefix.clone())
-                } else {
-                    self.replay_prefix_ops.take()
-                };
-                let inner_result = if i == 0 {
-                    self.parse_line_inner_from(replay_line, syntax_set, match_start_pos)
-                } else {
-                    self.parse_line_inner(replay_line, syntax_set)
-                };
-                self.replay_ctx = prev_replay_ctx;
-                self.replay_prefix_ops = prev_replay_prefix;
-                let tail_ops = inner_result?;
-                let line_ops = if i == 0 {
-                    let mut first_line_ops = first_line_prefix.clone();
-                    first_line_ops.extend(tail_ops);
-                    first_line_ops
-                } else {
-                    tail_ops
-                };
-                replayed_ops.push(line_ops);
-            }
-            // Capture inner corrections (if any), restore prior state, then
-            // prefer the inner corrections for overlapping indices. Without
-            // this, the outer's locally-computed `replayed_ops[i]` for
-            // indices an inner cross-line fail later corrected (during a
-            // later iteration of this same loop) would silently overwrite
-            // the inner's more accurate correction in `flushed_ops` —
-            // observed on Java's `@A.B\n(par=1)\nenum E {}` where the
-            // outer `declarations` cross-line replay's line-1 ops froze the
-            // dotted annotation as `path` alt before the inner
-            // `annotation-qualified-identifier` cross-line fail's `name`-alt
-            // resolution arrived (during line-2 reparse).
-            let mut inner_corrections = std::mem::take(&mut self.flushed_ops);
-            let mut inner_corrections_start = self.flushed_ops_start.take();
-            let mut inner_corrections_bp = std::mem::take(&mut self.flushed_ops_bp_per_slot);
-            self.flushed_ops = saved_flushed;
-            self.flushed_ops_start = saved_flushed_start;
-            self.flushed_ops_bp_per_slot = saved_flushed_bp;
-            let mut _inner_replay_max =
-                std::mem::take(&mut self.inner_replay_max_depth).unwrap_or_default();
-            self.inner_replay_max_depth = saved_max_depth;
-
-            // Composite-discriminator rewind: when the cross-line `class-members` BP
-            // exhausts an earlier alt without inner corrections and the inner replay reached
-            // a strictly-deeper BP under a `meta.function.return-type.java` push, force alt 5
-            // (`member-maybe-field`) by inlining its setup + replay loop here. The
-            // `next_alt_index < 5` short-circuit inside the discriminator prevents a re-fire
-            // after this block runs. The naive "set `next_alternative = 5` and let the caller
-            // re-trigger" variant doesn't work — the caller's re-parse doesn't re-encounter
-            // the same BP at the same byte position, so the target alt's replay never runs.
-            let return_type_scope =
-                Scope::new("meta.function.return-type.java").expect("well-formed scope");
-            if Self::class_members_alt5_should_rewind(
-                &outer_bp_info,
-                next_alt_index,
-                inner_corrections.is_empty(),
-                &_inner_replay_max,
-                &replayed_ops,
-                return_type_scope,
-            ) {
-                // Restore parser state to BP-creation snapshot. The
-                // `branch_points` snapshot has the BP at next_alternative
-                // = next_alt_index + 1 (post-line-2511 advance); we
-                // overwrite to 6 below to record alt 5 as just tried.
-                self.branch_points = branch_points_snapshot_for_rewind;
-                self.stack = stack_snapshot.clone();
-                self.proto_starts = proto_starts_snapshot.clone();
-                self.escape_stack = escape_stack_snapshot.clone();
-                self.first_line = first_line_snapshot;
-                *non_consuming_push_at = non_consuming_push_at_snapshot;
-
-                // Mirror line ~2511's pattern: advance past alt 5 so a
-                // future fail on this BP goes to the exhaustion path
-                // (alt 5 has been tried by the inline re-run below).
-                self.branch_points[bp_index].next_alternative = 5 + 1;
-
-                // Re-execute the alt-cycle setup for alt 5 (mirrors
-                // lines ~2495-2508 but with alt index 5 instead of N).
-                let next_alt_5 = self.branch_points[bp_index].alternatives[5].clone();
-                if pop_count > 0 {
-                    for _ in 0..pop_count {
-                        self.stack.pop();
-                    }
-                }
-                let with_prototype_5 = self.branch_points[bp_index].with_prototype.clone();
-                let context_id_5 = next_alt_5.id()?;
-                let proto_ids_5 = match with_prototype_5 {
-                    Some(ref p) => vec![p.id()?],
-                    None => Vec::new(),
-                };
-                self.stack.push(StateLevel {
-                    context: context_id_5,
-                    prototypes: proto_ids_5,
-                    captures: None,
-                });
-
-                // Re-build first_line_prefix for alt 5 (mirrors lines
-                // ~2611-2632).
-                let mut first_line_prefix_5 = prefix_ops.clone();
-                let synthetic_op_alt_5 = MatchOperation::Push {
-                    ctx_refs: vec![next_alt_5.clone()],
-                    pop_count,
-                };
-                let level_ctx_id_5 = stack_snapshot.last().map(|l| l.context);
-                let post_set_stack_5 = std::mem::replace(&mut self.stack, stack_snapshot.clone());
-                if let Some(level_ctx_id) = level_ctx_id_5 {
-                    let level_context = syntax_set.get_context(&level_ctx_id)?;
-                    self.push_meta_ops(
-                        true,
-                        trigger_match_start,
-                        level_context,
-                        &synthetic_op_alt_5,
-                        syntax_set,
-                        &mut first_line_prefix_5,
-                    )?;
-                    for scope in &trigger_pat_scope {
-                        first_line_prefix_5.push((trigger_match_start, ScopeStackOp::Push(*scope)));
-                    }
-                    first_line_prefix_5.extend(trigger_capture_ops.iter().cloned());
-                    if !trigger_pat_scope.is_empty() {
-                        first_line_prefix_5
-                            .push((match_start_pos, ScopeStackOp::Pop(trigger_pat_scope.len())));
-                    }
-                    self.push_meta_ops(
-                        false,
-                        match_start_pos,
-                        level_context,
-                        &synthetic_op_alt_5,
-                        syntax_set,
-                        &mut first_line_prefix_5,
-                    )?;
-                }
-                self.stack = post_set_stack_5;
-
-                // Re-run inner replay loop with alt 5 (mirrors lines
-                // ~2649-2693).
-                let saved_flushed_5 = std::mem::take(&mut self.flushed_ops);
-                let saved_flushed_start_5 = self.flushed_ops_start.take();
-                let saved_flushed_bp_5 = std::mem::take(&mut self.flushed_ops_bp_per_slot);
-                let saved_max_depth_5 =
-                    self.inner_replay_max_depth.replace(MaxDepthSeen::default());
-
-                let mut replayed_ops_5: Vec<Vec<(usize, ScopeStackOp)>> =
-                    Vec::with_capacity(truncated_lines.len());
-                for (i, replay_line) in truncated_lines.iter().enumerate() {
-                    let prev_replay_ctx = self.replay_ctx.replace(ReplayCtx {
-                        line_number: bp_line_number + i,
-                        pending_lines_snapshot_offset: pending_lines_snapshot_len + i,
-                    });
-                    let prev_replay_prefix = if i == 0 {
-                        self.replay_prefix_ops.replace(first_line_prefix_5.clone())
-                    } else {
-                        self.replay_prefix_ops.take()
-                    };
-                    let inner_result = if i == 0 {
-                        self.parse_line_inner_from(replay_line, syntax_set, match_start_pos)
-                    } else {
-                        self.parse_line_inner(replay_line, syntax_set)
-                    };
-                    self.replay_ctx = prev_replay_ctx;
-                    self.replay_prefix_ops = prev_replay_prefix;
-                    let tail_ops = inner_result?;
-                    let line_ops = if i == 0 {
-                        let mut first_line_ops = first_line_prefix_5.clone();
-                        first_line_ops.extend(tail_ops);
-                        first_line_ops
-                    } else {
-                        tail_ops
-                    };
-                    replayed_ops_5.push(line_ops);
-                }
-
-                // Capture and restore (mirrors lines ~2705-2713).
-                let inner_corrections_5 = std::mem::take(&mut self.flushed_ops);
-                let inner_corrections_start_5 = self.flushed_ops_start.take();
-                let inner_corrections_bp_5 = std::mem::take(&mut self.flushed_ops_bp_per_slot);
-                self.flushed_ops = saved_flushed_5;
-                self.flushed_ops_start = saved_flushed_start_5;
-                self.flushed_ops_bp_per_slot = saved_flushed_bp_5;
-                let inner_replay_max_5 =
-                    std::mem::take(&mut self.inner_replay_max_depth).unwrap_or_default();
-                self.inner_replay_max_depth = saved_max_depth_5;
-
-                // Reassign outer locals so the merge code below sees
-                // alt 5's results.
-                replayed_ops = replayed_ops_5;
-                inner_corrections = inner_corrections_5;
-                inner_corrections_start = inner_corrections_start_5;
-                inner_corrections_bp = inner_corrections_bp_5;
-                _inner_replay_max = inner_replay_max_5;
-            }
-
-            let mut replayed_per_slot_bp: Vec<BpInfo> =
-                vec![outer_bp_info.clone(); replayed_ops.len()];
-            if let Some(start) = inner_corrections_start {
-                if !inner_corrections.is_empty() {
-                    let inner_bp_first = inner_corrections_bp.first().cloned();
-                    self.prefer_inner_replay_corrections(
-                        pending_lines_snapshot_len,
-                        &mut replayed_ops,
-                        &mut replayed_per_slot_bp,
-                        &inner_corrections,
-                        start,
-                        &outer_bp_info,
-                        inner_bp_first.as_ref(),
-                        &inner_corrections_bp,
-                    );
-                }
-            }
-            self.merge_flushed(
-                pending_lines_snapshot_len,
-                replayed_ops,
-                replayed_per_slot_bp,
-            );
-
-            // Restart the current line from the beginning.
-            ops.clear();
-            *start = 0;
-            *non_consuming_push_at = (0, 0, 0);
-
-            // Guard: the replayed `parse_line_inner` calls above can
-            // mutate `self.branch_points` (adding new branches,
-            // removing expired or exhausted ones), which can shift or
-            // invalidate `bp_index`. Indexing with the stale position
-            // previously panicked outright on files that exercise
-            // nested cross-line branching (observed on
-            // `JavaScript/syntax_test_js.js` and
-            // `syntax_test_typescript.ts`). Skip the bookkeeping if
-            // the branch point has been removed — the replay already
-            // completed, which is the essential work of the fail.
-            if bp_index < self.branch_points.len() {
-                self.branch_points[bp_index].ops_snapshot_len = 0;
-            }
-        } else {
-            // Same-line fail: truncate ops back to the snapshot point and rewind.
-            ops.truncate(ops_snapshot_len.min(ops.len()));
-            // Empty-line continuation: when the same-line fail fires at
-            // pos 0 of an empty line (just `\n`), the alt-N replacement
-            // is typically a `match: '' pop: N` (e.g. Markdown's
-            // `link-def-attr-continuation` failing into
-            // `immediately-pop2`). Executing that pop at pos 0 emits
-            // visible scope Pops on the empty line's only character —
-            // collapsing the parent meta_scope (e.g.
-            // `meta.link.reference.def.markdown`) before the empty
-            // line's own scope is recorded. Advance to past-EOL so the
-            // pop emits there instead, which ScopeRegionIterator wraps
-            // to the next line's baseline. ST's behavior matches: the
-            // LRD pop straddles line 3→line 4, not line 2→line 3.
-            let resume = if match_start_pos == 0 && line.len() <= 1 && line.trim().is_empty() {
-                line.len()
-            } else {
-                match_start_pos
-            };
-            *start = resume;
-
-            // Keep `ops_snapshot_len` pointing at the pre-branch state.
-            // Subsequent fails on the same branch_point must truncate
-            // back to *here* — not to the position after the pat.scope
-            // re-emit below — otherwise a second fail would preserve
-            // the first re-emit's (Push at trigger_match_start) while
-            // appending another re-emit, producing the disordered
-            // sequence (trigger, Push), (match_end, Pop), (trigger,
-            // Push), (match_end, Pop). `ScopeRegionIterator` then
-            // panics in `easy.rs` because position goes backwards.
-            self.branch_points[bp_index].ops_snapshot_len = ops.len();
-
-            // Re-emit the branch_point match's own scopes over their
-            // original span. Without this, keywords that trigger a
-            // branch (e.g. `LIKE` with
-            // `scope: keyword.operator.comparison.sql`) lose their
-            // scope whenever alt[0] fails and alt[1..] succeeds,
-            // because the original Push/Pop pair was truncated off
-            // `ops` together with alt[0]'s subsequent work.
-            //
-            // Use `push_meta_ops` with a synthetic Set/Push for the
-            // new alternative — same path the original branch creation
-            // takes — so both the new alternative's own meta scopes
-            // AND the popped contexts' meta_scope/mcs clearance Pop
-            // (for `pop: N + branch_point`, N > 0) get re-emitted.
-            // A bespoke re-emit of just `context.meta_scope` /
-            // `context.meta_content_scope` was missing the
-            // popped-contexts Pop, leaving Java's
-            // `pop: 2 + branch_point: annotation-qualified-parameters`
-            // with `meta.annotation.identifier.java meta.path.java`
-            // (annotation-qualified-identifier's `meta_scope`) leaked
-            // on the stack after the branch_point's first alt failed
-            // and the second alt (`immediately-pop`) ran.
-            //
-            // `push_meta_ops` reads `self.stack` to compute the
-            // popped contexts' scope atoms, so swap in `stack_snapshot`
-            // (pre-pop state captured at branch creation) for the
-            // duration of the calls — `self.stack` currently holds the
-            // post-set state (alt N already pushed).
-            let synthetic_op_alt_n = MatchOperation::Push {
-                ctx_refs: vec![next_alt.clone()],
-                pop_count,
-            };
-            let level_ctx_id = stack_snapshot.last().map(|l| l.context);
-            let post_set_stack = std::mem::replace(&mut self.stack, stack_snapshot.clone());
-            if let Some(level_ctx_id) = level_ctx_id {
-                let level_context = syntax_set.get_context(&level_ctx_id)?;
-                self.push_meta_ops(
-                    true,
-                    trigger_match_start,
-                    level_context,
-                    &synthetic_op_alt_n,
-                    syntax_set,
-                    ops,
-                )?;
-                for scope in &trigger_pat_scope {
-                    ops.push((trigger_match_start, ScopeStackOp::Push(*scope)));
-                }
-                // Captures emitted alongside the original pat.scope (e.g.
-                // `keyword.declaration.data.haskell` on the first capture of
-                // `(data)(?:\s+(family|instance))?`) were truncated off with
-                // alt[0]'s ops. Re-emit them inside the pat_scope brackets so
-                // the keyword scope survives the branch swap.
-                ops.extend(trigger_capture_ops.iter().cloned());
-                if !trigger_pat_scope.is_empty() {
-                    ops.push((match_start_pos, ScopeStackOp::Pop(trigger_pat_scope.len())));
-                }
-                self.push_meta_ops(
-                    false,
-                    match_start_pos,
-                    level_context,
-                    &synthetic_op_alt_n,
-                    syntax_set,
-                    ops,
-                )?;
-            }
-            self.stack = post_set_stack;
-        }
-
-        // Clear search cache since we're rewinding.
-        search_cache.clear();
-
-        Ok(true)
-    }
-
-    /// Get the syntax version for the current parse state
-    fn current_syntax_version(&self, syntax_set: &SyntaxSet) -> u32 {
-        if let Some(level) = self.stack.last() {
-            let syntax_index = level.context.syntax_index;
-            syntax_set
-                .syntaxes()
-                .get(syntax_index)
-                .map_or(1, |s| s.version)
-        } else {
-            1
-        }
-    }
-
-    fn push_meta_ops(
-        &self,
-        initial: bool,
-        index: usize,
-        cur_context: &Context,
-        match_op: &MatchOperation,
-        syntax_set: &SyntaxSet,
-        ops: &mut Vec<(usize, ScopeStackOp)>,
-    ) -> Result<(), ParsingError> {
-        let version = self.current_syntax_version(syntax_set);
-        // println!("metas ops for {:?}, initial: {}",
-        //          match_op,
-        //          initial);
-        // println!("{:?}", cur_context.meta_scope);
-        match *match_op {
-            MatchOperation::Pop(n) => {
-                // For `pop: N` with N > 1, every context being popped
-                // contributes scope atoms on the scope stack that must
-                // be unwound in LIFO order. The TOP context's trigger
-                // text must not see its own `meta_content_scope`, so
-                // that one is popped in the initial phase; all other
-                // scope unwinding (top context's `meta_scope`, then
-                // each deeper context's `meta_content_scope` followed
-                // by its `meta_scope`) happens in the non-initial
-                // phase, immediately after the match text's own
-                // scope has been popped.
-                //
-                // Before this fix only the top context's scopes were
-                // ever popped, leaving the N-1 deeper contexts'
-                // `meta_scope` / `meta_content_scope` atoms orphaned —
-                // the cause of the "scope stack grows unboundedly"
-                // cascade in Makefile and Zsh (Category A).
-                let stack_len = self.stack.len();
-                let pop_count = n.min(stack_len);
-                if initial {
-                    // v2: if the context immediately below the top has
-                    // embed_scope_replaces, cur_context's meta_content_scope
-                    // was never pushed, so don't generate a Pop for it.
-                    let skip = version >= 2
-                        && stack_len >= 2
-                        && syntax_set
-                            .get_context(&self.stack[stack_len - 2].context)
-                            .map(|c| c.embed_scope_replaces)
-                            .unwrap_or(false);
-                    if !skip && !cur_context.meta_content_scope.is_empty() {
-                        ops.push((
-                            index,
-                            ScopeStackOp::Pop(cur_context.meta_content_scope.len()),
-                        ));
-                    }
-                } else {
-                    // Top context's meta_scope comes off first (it sat
-                    // immediately below the trigger text's scope on the
-                    // stack).
-                    if !cur_context.meta_scope.is_empty() {
-                        ops.push((index, ScopeStackOp::Pop(cur_context.meta_scope.len())));
-                    }
-                    // Restore cur_context's `clear_scopes` BEFORE
-                    // popping deeper contexts. The Clear hid the
-                    // deeper context's meta_scope/mcs atoms; with the
-                    // Restore deferred to the very end, the
-                    // depth-loop's `Pop(deep.meta_scope.len())` would
-                    // pop visible-stack scopes that don't belong to
-                    // the deeper context — observed on Java's
-                    // `case DayType when -> "incomplete"`, where
-                    // `case-label-expression`'s `clear_scopes: 1`
-                    // cleared `case-label`'s `meta.case.java` and
-                    // `case-label-end`'s `pop: 2` then popped the
-                    // surrounding `meta.block.java` (switch's block)
-                    // off the consumer's stack instead.
-                    if cur_context.clear_scopes.is_some() {
-                        ops.push((index, ScopeStackOp::Restore))
-                    }
-                    // Each deeper context's scopes are popped in
-                    // top-to-bottom order: meta_content_scope first
-                    // (pushed after its own meta_scope, hence above on
-                    // the stack), then meta_scope, then any Restore
-                    // paired with that context's own `clear_scopes`
-                    // (mirrors the push order Clear → meta_scope → mcs
-                    // in reverse).
-                    for depth in 1..pop_count {
-                        let level_idx = stack_len - 1 - depth;
-                        let ctx = syntax_set.get_context(&self.stack[level_idx].context)?;
-                        let skip_content = version >= 2
-                            && level_idx >= 1
-                            && syntax_set
-                                .get_context(&self.stack[level_idx - 1].context)
-                                .map(|c| c.embed_scope_replaces)
-                                .unwrap_or(false);
-                        if !skip_content && !ctx.meta_content_scope.is_empty() {
-                            ops.push((index, ScopeStackOp::Pop(ctx.meta_content_scope.len())));
-                        }
-                        if !ctx.meta_scope.is_empty() {
-                            ops.push((index, ScopeStackOp::Pop(ctx.meta_scope.len())));
-                        }
-                        if ctx.clear_scopes.is_some() {
-                            ops.push((index, ScopeStackOp::Restore));
-                        }
-                    }
-                }
-            }
-            // for some reason the ST3 behaviour of set is convoluted and is inconsistent with the docs and other ops
-            // - the meta_content_scope of the current context is applied to the matched thing, unlike pop
-            // - the clear_scopes are applied after the matched token, unlike push
-            // - the interaction with meta scopes means that the token has the meta scopes of both the current scope and the new scope.
-            MatchOperation::Push {
-                ctx_refs: ref context_refs,
-                ..
-            }
-            | MatchOperation::Set {
-                ctx_refs: ref context_refs,
-                ..
-            } => {
-                let is_set = matches!(*match_op, MatchOperation::Set { .. });
-                let set_pop_count = match *match_op {
-                    MatchOperation::Set { pop_count, .. } => pop_count.max(1),
-                    _ => 1,
-                };
-                // `pop: N + push:` is **lookahead**: per ST docs, when pop is
-                // combined with push the pop happens first and the matched
-                // text is treated as a lookahead, so the popped frames'
-                // meta_scope atoms must be dropped from the visible stack
-                // before the trigger token records its scope. (`pop: N + set:`
-                // is the documented exception — stacking — handled below.)
-                let push_pop_count = match *match_op {
-                    MatchOperation::Push { pop_count, .. } => pop_count,
-                    _ => 0,
-                };
-                let is_push_with_pop = push_pop_count > 0;
-                // a match pattern that "set"s keeps the meta_content_scope and meta_scope from the previous context
-                if initial {
-                    // v2: pop the USER-DECLARED part of cur.mcs so the
-                    // matched text doesn't see it. The AUTO-INJECTED
-                    // top-level scope (added to `main.meta_content_scope[0]`
-                    // by `add_initial_contexts`) stays on the stack across
-                    // the trigger — verified against ST 4200 stable on
-                    // TOML's `[section]` rule, where the `[` trigger sees
-                    // `source.toml` (TOML main's auto-injected mcs)
-                    // alongside the trigger's own `meta.section.toml`. The
-                    // distinction matches ST's documented v2 set behavior:
-                    // the matched text doesn't inherit cur.mcs, but the
-                    // file's top-level scope is conceptually always on.
-                    //
-                    // Skip when cur_context's mcs was never pushed because
-                    // the context immediately below has
-                    // `embed_scope_replaces` (the embedded syntax's main
-                    // mcs is suppressed in favor of `embed_scope` on the
-                    // wrapper). Without this, the Pop takes off the
-                    // topmost wrapper-pushed scope — observed on Markdown
-                    // bash fenced blocks where `source.shell.bash` (the
-                    // last embed_scope token) was disappearing on the
-                    // embedded main's first `set:` rule.
-                    let stack_len = self.stack.len();
-                    let skip_cur_mcs_pop = version >= 2
-                        && stack_len >= 2
-                        && syntax_set
-                            .get_context(&self.stack[stack_len - 2].context)
-                            .map(|c| c.embed_scope_replaces)
-                            .unwrap_or(false);
-                    if is_set
-                        && version >= 2
-                        && !cur_context.meta_content_scope.is_empty()
-                        && !skip_cur_mcs_pop
-                    {
-                        // Identify the auto-injected top-level scope: the
-                        // syntax's `scope:` directive lands at
-                        // `main.meta_content_scope[0]` via
-                        // `add_initial_contexts`. Compare cur.mcs[0] to
-                        // the syntax's top-level scope; if they match,
-                        // exclude position 0 from the Pop so the matched
-                        // text retains the file scope.
-                        let cur_syntax_idx = self.stack[stack_len - 1].context.syntax_index;
-                        let top_level_scope =
-                            syntax_set.syntaxes().get(cur_syntax_idx).map(|s| s.scope);
-                        let mut pop_count = cur_context.meta_content_scope.len();
-                        if top_level_scope == cur_context.meta_content_scope.first().copied() {
-                            pop_count -= 1;
-                        }
-                        if pop_count > 0 {
-                            ops.push((index, ScopeStackOp::Pop(pop_count)));
-                        }
-                    }
-                    // `pop: N + push:` with N > 1: per ST's lookahead rule,
-                    // the (N-1) deeper popped contexts' meta_scope /
-                    // meta_content_scope atoms must be dropped from the
-                    // visible stack BEFORE the trigger token records its
-                    // scope. Without this, the trigger token observes the
-                    // leaked deeper meta_scope — observed on Java's
-                    // `@RunWith(JUnit4.class)` where
-                    // `pop: 2 + push: annotation-parameters-body` left
-                    // `meta.annotation.identifier.java`
-                    // (annotation-unqualified-identifier's meta_scope) on
-                    // the stack at the `(` token instead of just
-                    // `meta.annotation.parameters.java meta.group.java`.
-                    //
-                    // `pop: N + set:` is the ST-documented exception
-                    // (stacking, not lookahead): the trigger token
-                    // RECEIVES the popped frames' meta_scope. That path is
-                    // handled in the non-initial phase below — both the
-                    // genuine stacking case (target has its own ms) and the
-                    // case where the target has no ms (e.g. TS's
-                    // `(?:get|set|async){{identifier_break}} pop: 2 + set:`
-                    // in `object-property-name`, where ST keeps
-                    // `meta.mapping.key.js` visible to the `get` token).
-                    //
-                    // Skip when cur or any popped deeper context has
-                    // `clear_scopes`: those interact with the clear_stack
-                    // through the existing non-initial pipeline (Restore
-                    // ordering vs head_pop, multi-target Clear preview),
-                    // and the rotation here would race with that.
-                    let mut apply_initial_deeper_pop = is_push_with_pop && push_pop_count > 1;
-                    apply_initial_deeper_pop &= cur_context.clear_scopes.is_none();
-                    if apply_initial_deeper_pop {
-                        let stack_len = self.stack.len();
-                        for depth in 1..push_pop_count.min(stack_len) {
-                            let level_idx = stack_len - 1 - depth;
-                            let ctx = syntax_set.get_context(&self.stack[level_idx].context)?;
-                            if ctx.clear_scopes.is_some() {
-                                apply_initial_deeper_pop = false;
-                                break;
-                            }
-                        }
-                    }
-                    if apply_initial_deeper_pop {
-                        let cur_ms_rotate = cur_context.meta_scope.len();
-                        if cur_ms_rotate > 0 {
-                            ops.push((index, ScopeStackOp::Pop(cur_ms_rotate)));
-                        }
-                        let stack_len = self.stack.len();
-                        for depth in 1..push_pop_count.min(stack_len) {
-                            let level_idx = stack_len - 1 - depth;
-                            let ctx = syntax_set.get_context(&self.stack[level_idx].context)?;
-                            if !ctx.meta_content_scope.is_empty() {
-                                ops.push((index, ScopeStackOp::Pop(ctx.meta_content_scope.len())));
-                            }
-                            if !ctx.meta_scope.is_empty() {
-                                ops.push((index, ScopeStackOp::Pop(ctx.meta_scope.len())));
-                            }
-                        }
-                        for scope in cur_context.meta_scope.iter() {
-                            ops.push((index, ScopeStackOp::Push(*scope)));
-                        }
-                    }
-                    // NOTE: cur_context.clear_scopes Restore is emitted in the
-                    // non-initial phase below, AFTER Pop(cur.meta_scope + target.meta_scope)
-                    // has run. Restoring here (pre-match) would place the cleared
-                    // scopes on top of the stack above the target.meta_scope push,
-                    // and the non-initial Pop would then remove the restored scopes
-                    // instead of the intended meta_scopes — dropping cur's cleared
-                    // state on the floor. Observed as duplicate
-                    // `meta.mapping.value.json` atoms in nested JSON objects.
-                    // add each context's meta scope
-                    if version >= 2 {
-                        // Push: emit Clear for every pushed context that has
-                        // `clear_scopes`, at its own index position in the
-                        // push order — same as v1. Sublime permits at most
-                        // one `clear_scopes` per push list, but when it sits
-                        // on a non-topmost entry (e.g. Python's
-                        // `f-string-replacement-meta` at index 0 of a
-                        // 3-context push) restricting to `i == last_idx`
-                        // silently drops it, leaking the parent's
-                        // `meta_content_scope` atoms into interpolation
-                        // content.
-                        //
-                        // Single-context `set:` with clear_scopes on the
-                        // target: emit Clear here — before target.meta_scope
-                        // is pushed and before the trigger match scope — so
-                        // the matched text sees the cleared stack.
-                        // Observed on Lisp's `(defun fn (...)`: the
-                        // parameter-list `(` otherwise kept the enclosing
-                        // `meta.function.lisp` alongside
-                        // `meta.function.parameters.lisp` because Clear
-                        // fired only after the match in the non-initial
-                        // phase. See
-                        // `v2_set_to_target_with_clear_scopes_clears_parent_meta_content_scope`.
-                        //
-                        // Exception: when cur_context itself carries
-                        // `meta_scope` / `meta_content_scope`, those atoms sit
-                        // on top of the visible stack at this point and the
-                        // initial-phase Clear would hide them. The non-initial
-                        // Pop (sized by cur.ms.len() + target.ms.len()) would
-                        // then pop the wrong atoms (from below cur's ms),
-                        // and the Restore that follows would resurrect cur's
-                        // ms back onto the stack instead of the parent atoms
-                        // the Clear was meant to hide. Defer the Clear into
-                        // the non-initial phase (after Pop+Restore) so it
-                        // hides parent atoms, not cur's. Bash's
-                        // `tilde-modifier` (clear+ms) → set:
-                        // `tilde-modifier-username` (clear+mcs) is the
-                        // canonical instance. See
-                        // `cur_meta_scope_set_to_target_with_clear_scopes`.
-                        //
-                        // Multi-context `set:` keeps Clear in the non-initial
-                        // phase (emitted inline after preceding contexts'
-                        // mcs pushes). Moving it to the initial phase here
-                        // would strip atoms from below the outer mcs rather
-                        // than from the top of the just-pushed inner mcs
-                        // stack — Makefile's `set: [value-to-be-defined,
-                        // eat-whitespace-then-pop]` relies on Clear eating
-                        // the last-pushed mcs atom, which Restore then
-                        // replaces when eat-whitespace-then-pop pops.
-                        let cur_has_meta = !cur_context.meta_scope.is_empty()
-                            || !cur_context.meta_content_scope.is_empty();
-                        let single_context_set_clear =
-                            is_set && context_refs.len() == 1 && !cur_has_meta;
-                        // Multi-context `set:` whose target body declares
-                        // `clear_scopes: N` AND a non-empty `meta_scope`,
-                        // with cur empty: ST drops one EXTRA atom beyond
-                        // Clear(N) on the trigger token. The body content
-                        // sees only Clear(N) atoms gone. Observed on PHP
-                        // `function bye(): never {` — at the `:`, ST drops
-                        // both `meta.function.php` (function-block's mcs,
-                        // what Clear(1) would clear) AND the next-deeper
-                        // `source.php.embedded.html` (the embed wrapper's
-                        // mcs); syntect previously kept both, leaking
-                        // nested `meta.function.php` /
-                        // `meta.function.return-type.php` into the colon.
-                        // See `php_multi_set_target_clear_drops_extra_parent_mcs_on_trigger`.
-                        //
-                        // The target's `meta_scope` non-emptiness is what
-                        // anchors the extra drop on the trigger: that ms
-                        // is pushed on top of the trigger and asks ST to
-                        // strip one more parent atom below Clear's reach.
-                        // Targets with `meta_content_scope` only (e.g.
-                        // Zsh's `zsh-redirection-glob-range-end`, which has
-                        // `clear_scopes: 1` + `meta_content_scope` but no
-                        // `meta_scope`) must NOT trigger this — there's no
-                        // ms to anchor the extra drop on the trigger token,
-                        // so doing it strips fundamental scopes
-                        // (`source.shell.zsh`,
-                        // `meta.function-call.arguments.shell`) that ST
-                        // keeps.
-                        let target_clear_amt = if is_set {
-                            context_refs.iter().find_map(|r| {
-                                r.resolve(syntax_set)
-                                    .ok()
-                                    .and_then(|c| match c.clear_scopes {
-                                        Some(ClearAmount::TopN(n))
-                                            if n > 0 && !c.meta_scope.is_empty() =>
-                                        {
-                                            Some(n)
-                                        }
-                                        _ => None,
-                                    })
-                            })
-                        } else {
-                            None
-                        };
-                        let cur_inert = !cur_has_meta && cur_context.clear_scopes.is_none();
-                        let multi_set_extra_drop = is_set
-                            && set_pop_count == 1
-                            && context_refs.len() > 1
-                            && cur_inert
-                            && target_clear_amt.is_some();
-                        if let (true, Some(amt)) = (multi_set_extra_drop, target_clear_amt) {
-                            ops.push((index, ScopeStackOp::Clear(ClearAmount::TopN(amt + 1))));
-                        }
-                        // Multi-context `set:` whose non-topmost target has
-                        // `clear_scopes: N` + an empty `meta_scope`
-                        // (`meta_content_scope`-only): ST applies the Clear
-                        // to atoms that EARLIER targets pushed via their
-                        // `meta_scope`, and the strip is visible to the
-                        // trigger token's own scopes — observed on Zsh's
-                        // `zsh-redirection-glob-range-begin`'s
-                        //   set: [string-path-pattern-body,
-                        //         zsh-redirection-glob-range-end, …]
-                        // where `string-path-pattern-body` pushes
-                        // `meta.string.glob.shell string.unquoted.shell` and
-                        // `…-range-end` declares `clear_scopes: 1`. The
-                        // capture-2 scope `meta.range.shell.zsh
-                        // punctuation.definition.range.begin.shell.zsh` on
-                        // the `<` is asserted with `- string`, so
-                        // `string.unquoted.shell` must be hidden at the
-                        // trigger. Without this preview Clear it leaks into
-                        // every glob-range opening. The matching Restore is
-                        // emitted at the start of the non-initial phase
-                        // below so `head_pop` finds the full visible stack.
-                        // See
-                        // `v2_multi_set_non_topmost_clear_scopes_strips_preceding_meta_scope_at_trigger`.
-                        let mut initial_atoms_pushed: usize = 0;
-                        for r in context_refs.iter() {
-                            let ctx = r.resolve(syntax_set)?;
-
-                            if is_set && context_refs.len() > 1 && ctx.meta_scope.is_empty() {
-                                if let Some(ClearAmount::TopN(n)) = ctx.clear_scopes {
-                                    let initial_clear = n.min(initial_atoms_pushed);
-                                    if initial_clear > 0 {
-                                        ops.push((
-                                            index,
-                                            ScopeStackOp::Clear(ClearAmount::TopN(initial_clear)),
-                                        ));
-                                        initial_atoms_pushed -= initial_clear;
-                                    }
-                                }
-                            }
-
-                            let emit_clear_here = !is_set || single_context_set_clear;
-                            if emit_clear_here {
-                                if let Some(clear_amount) = ctx.clear_scopes {
-                                    ops.push((index, ScopeStackOp::Clear(clear_amount)));
-                                }
-                            }
-
-                            for scope in ctx.meta_scope.iter() {
-                                ops.push((index, ScopeStackOp::Push(*scope)));
-                                initial_atoms_pushed += 1;
-                            }
-                        }
-                    } else {
-                        for r in context_refs.iter() {
-                            let ctx = r.resolve(syntax_set)?;
-
-                            if !is_set {
-                                if let Some(clear_amount) = ctx.clear_scopes {
-                                    ops.push((index, ScopeStackOp::Clear(clear_amount)));
-                                }
-                            }
-
-                            for scope in ctx.meta_scope.iter() {
-                                ops.push((index, ScopeStackOp::Push(*scope)));
-                            }
-                        }
-                    }
-                } else {
-                    // `pop: N + set:` (set_pop_count > 1) unwinds N-1 deeper
-                    // contexts in addition to the usual set-replace semantics;
-                    // their meta_scope / meta_content_scope atoms sitting on
-                    // the scope stack must be popped off, so force repush to
-                    // fire even if the immediate contexts had no mcs/ms.
-                    let repush = (is_set
-                        && (set_pop_count > 1
-                            || !cur_context.meta_scope.is_empty()
-                            || !cur_context.meta_content_scope.is_empty()
-                            // cur has clear_scopes but no meta_scope/mcs: we still
-                            // need to Pop the target.meta_scope pushed in initial,
-                            // Restore cur.clear_scopes, and re-push target.meta_scope
-                            // + target.meta_content_scope in the correct order.
-                            || cur_context.clear_scopes.is_some()))
-                        || context_refs.iter().any(|r| {
-                            let ctx = r.resolve(syntax_set).unwrap();
-
-                            !ctx.meta_content_scope.is_empty()
-                                || (ctx.clear_scopes.is_some() && is_set)
-                        });
-                    if repush {
-                        // Head pop: target.meta_scope (just pushed in the
-                        // initial phase) + cur's own meta scopes. These come
-                        // off as one Pop because they sit at the top of the
-                        // visible stack and don't need per-frame Restore.
-                        let target_ms_sum: usize = context_refs
-                            .iter()
-                            .map(|r| {
-                                let ctx = r.resolve(syntax_set).unwrap();
-                                ctx.meta_scope.len()
-                            })
-                            .sum();
-                        let mut head_pop = target_ms_sum;
-                        if is_set {
-                            if version >= 2 {
-                                // v2: the user-declared part of
-                                // cur.meta_content_scope was popped in the
-                                // initial phase (the auto-injected
-                                // top-level scope, if any, stays). Only
-                                // cur.meta_scope sits on top of the
-                                // visible stack at this point.
-                                head_pop += cur_context.meta_scope.len();
-                            } else {
-                                head_pop += cur_context.meta_content_scope.len()
-                                    + cur_context.meta_scope.len();
-                            }
-                        }
-
-                        // `pop: N + set:` with clear_scopes on the leaving
-                        // context: restore the cleared atoms BEFORE the
-                        // head Pop so its count finds the visible stack
-                        // intact. Without this, Pop eats atoms from below
-                        // the popped range — observed on Batch File
-                        // `cmd-set-quoted-value-inner-end` (`clear_scopes: 1`)
-                        // firing `pop: 2, set: ignored-tail-outer`, which
-                        // otherwise drops `meta.command.set.dosbatch` from
-                        // the trailing content of every `set "var"=...`
-                        // line.
-                        let restore_before_pop =
-                            is_set && set_pop_count > 1 && cur_context.clear_scopes.is_some();
-                        if restore_before_pop {
-                            ops.push((index, ScopeStackOp::Restore));
-                        }
-
-                        // Pair to the initial-phase preview Clears emitted
-                        // above for non-topmost `clear_scopes` targets with
-                        // empty `meta_scope`. Each Restore here undoes one
-                        // such preview Clear so the upcoming `head_pop`
-                        // sees the full visible stack the captures pushed
-                        // onto. The body's matching Clears are re-emitted
-                        // below in the per-target loop, so the steady-state
-                        // post-trigger view is unchanged.
-                        if is_set && version >= 2 && context_refs.len() > 1 {
-                            let mut atoms = 0usize;
-                            let mut preview_clears = 0usize;
-                            for r in context_refs.iter() {
-                                let ctx = r.resolve(syntax_set)?;
-                                if ctx.meta_scope.is_empty() {
-                                    if let Some(ClearAmount::TopN(n)) = ctx.clear_scopes {
-                                        let initial_clear = n.min(atoms);
-                                        if initial_clear > 0 {
-                                            preview_clears += 1;
-                                            atoms -= initial_clear;
-                                        }
-                                    }
-                                }
-                                atoms += ctx.meta_scope.len();
-                            }
-                            for _ in 0..preview_clears {
-                                ops.push((index, ScopeStackOp::Restore));
-                            }
-                        }
-
-                        if head_pop > 0 {
-                            ops.push((index, ScopeStackOp::Pop(head_pop)));
-                        }
-
-                        // Restore scopes cleared by the leaving context, now that
-                        // cur.meta_scope and the initial phase's target.meta_scope
-                        // push have been popped off. The restored atoms land below
-                        // the target's upcoming meta_scope / meta_content_scope push.
-                        if is_set && cur_context.clear_scopes.is_some() && !restore_before_pop {
-                            ops.push((index, ScopeStackOp::Restore));
-                        }
-
-                        // `pop: N + set:` (set_pop_count > 1) is **stacking**
-                        // per ST docs: the trigger token receives the popped
-                        // frames' meta_scope, and the deeper-pop happens
-                        // AFTER the trigger records its scope. Mirror the
-                        // `MatchOperation::Pop` arm: pop each deeper frame's
-                        // mcs+ms in top-to-bottom order, then Restore that
-                        // frame's `clear_scopes` if any. Without the
-                        // per-depth Restore, atoms cleared by a deeper
-                        // frame stay in clear_stack out of reach, and the
-                        // per-target Clear below then bites one atom too
-                        // deep. Observed on Python regex inside a
-                        // `r'''(?ix:...)` triple-quoted string: the
-                        // activate-x-mode `pop: 3 + set:[...]` left the
-                        // outer's cleared `meta.mode.extended.regexp` in
-                        // clear_stack; the inner's `clear_scopes: 1` then
-                        // cleared `source.regexp.python` instead.
-                        //
-                        // Always emits for Set-with-pop. The Push-with-pop
-                        // (lookahead) path popped the deeper frames in the
-                        // initial phase already; it doesn't reach this arm
-                        // because `is_set` is false there.
-                        if is_set && set_pop_count > 1 {
-                            let stack_len = self.stack.len();
-                            for depth in 1..set_pop_count.min(stack_len) {
-                                let level_idx = stack_len - 1 - depth;
-                                let ctx = syntax_set.get_context(&self.stack[level_idx].context)?;
-                                if !ctx.meta_content_scope.is_empty() {
-                                    ops.push((
-                                        index,
-                                        ScopeStackOp::Pop(ctx.meta_content_scope.len()),
-                                    ));
-                                }
-                                if !ctx.meta_scope.is_empty() {
-                                    ops.push((index, ScopeStackOp::Pop(ctx.meta_scope.len())));
-                                }
-                                if ctx.clear_scopes.is_some() {
-                                    ops.push((index, ScopeStackOp::Restore));
-                                }
-                            }
-                        }
-
-                        // Pair to the initial-phase Clear(N+1) emitted for the
-                        // multi-context-set + cur-empty + target-clear case
-                        // above. The body content needs only the target's own
-                        // Clear(N) applied (emitted by the per-context loop
-                        // below); restoring here brings the (N+1)-atom batch
-                        // back onto the live stack, then the per-context
-                        // Clear(N) eats N of them and leaves the extra atom
-                        // visible to the body. The target context's matching
-                        // Restore in `MatchOperation::Pop` then unwinds the
-                        // smaller batch and the larger trigger-only batch is
-                        // left consumed.
-                        let cur_inert_restore = !cur_context.meta_scope.is_empty()
-                            || !cur_context.meta_content_scope.is_empty()
-                            || cur_context.clear_scopes.is_some();
-                        let target_clear_amt_restore = if is_set {
-                            context_refs.iter().find_map(|r| {
-                                r.resolve(syntax_set)
-                                    .ok()
-                                    .and_then(|c| match c.clear_scopes {
-                                        Some(ClearAmount::TopN(n))
-                                            if n > 0 && !c.meta_scope.is_empty() =>
-                                        {
-                                            Some(n)
-                                        }
-                                        _ => None,
-                                    })
-                            })
-                        } else {
-                            None
-                        };
-                        let multi_set_extra_drop_restore = is_set
-                            && version >= 2
-                            && set_pop_count == 1
-                            && context_refs.len() > 1
-                            && !cur_inert_restore
-                            && target_clear_amt_restore.is_some();
-                        if multi_set_extra_drop_restore {
-                            ops.push((index, ScopeStackOp::Restore));
-                        }
-
-                        // now we push meta scope and meta context scope for each context pushed
-                        if version >= 2 {
-                            // v2: For multi-context `set:`, Clear is emitted
-                            // here so it strips the topmost just-pushed mcs
-                            // atom (as Sublime does for multi-context set).
-                            // Single-context `set:` ordinarily emits Clear in
-                            // the initial phase (so the trigger token sees
-                            // the cleared stack); re-emitting here would
-                            // double-push onto clear_stack and cause Pop
-                            // underflow when the context unwinds.
-                            //
-                            // Exception: when cur_context has its own
-                            // `meta_scope` / `meta_content_scope` the initial
-                            // phase deferred the Clear to here so the Pop
-                            // above could find cur's ms on the visible stack.
-                            // Re-introduce the Clear now, after Pop+Restore,
-                            // so it strips parent atoms (the intended
-                            // target) rather than cur's ms.
-                            //
-                            // Clear is emitted per-context (not only for the
-                            // topmost) because `clear_scopes` on a non-topmost
-                            // context is a real pattern: Bash's
-                            // `set: [def-function-body, def-function-params,
-                            // def-function-name]` has `clear_scopes: 1` on
-                            // def-function-params (middle). Each Clear is
-                            // placed just before that context's own mcs/ms
-                            // pushes so it strips the previous iteration's
-                            // last-pushed atom, matching Sublime's semantics.
-                            let cur_has_meta = !cur_context.meta_scope.is_empty()
-                                || !cur_context.meta_content_scope.is_empty();
-                            let single_context_set_clear =
-                                is_set && context_refs.len() == 1 && !cur_has_meta;
-                            let mut prev_embed_scope_replaces = false;
-                            for r in context_refs.iter() {
-                                let ctx = r.resolve(syntax_set)?;
-
-                                if is_set && !single_context_set_clear {
-                                    if let Some(clear_amount) = ctx.clear_scopes {
-                                        ops.push((index, ScopeStackOp::Clear(clear_amount)));
-                                    }
-                                }
-
-                                for scope in ctx.meta_scope.iter() {
-                                    ops.push((index, ScopeStackOp::Push(*scope)));
-                                }
-                                // v2: if the previous context has embed_scope_replaces,
-                                // skip this context's meta_content_scope (the embedded
-                                // syntax's top-level scope is replaced by embed_scope)
-                                if !prev_embed_scope_replaces {
-                                    for scope in ctx.meta_content_scope.iter() {
-                                        ops.push((index, ScopeStackOp::Push(*scope)));
-                                    }
-                                }
-                                prev_embed_scope_replaces = ctx.embed_scope_replaces;
-                            }
-                        } else {
-                            for r in context_refs {
-                                let ctx = r.resolve(syntax_set)?;
-
-                                // for some reason, contrary to my reading of the docs, set does this after the token
-                                if is_set {
-                                    if let Some(clear_amount) = ctx.clear_scopes {
-                                        ops.push((index, ScopeStackOp::Clear(clear_amount)));
-                                    }
-                                }
-
-                                for scope in ctx.meta_scope.iter() {
-                                    ops.push((index, ScopeStackOp::Push(*scope)));
-                                }
-                                for scope in ctx.meta_content_scope.iter() {
-                                    ops.push((index, ScopeStackOp::Push(*scope)));
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            MatchOperation::Embed {
-                ref contexts,
-                pop_count,
-                ..
-            } => {
-                // `pop: N + embed:` is **lookahead** per ST: the trigger
-                // token must NOT inherit the popped frames' meta_scope.
-                // Route through `Push { pop_count }` so the existing
-                // Push-with-pop lookahead path in `push_meta_ops` /
-                // `perform_op` handles deeper-pop emission. The
-                // pre-recursive Pops below cover the embed-specific
-                // cur_context meta_scope suppression (depth 0); the
-                // Push-with-pop deeper-pop loop covers depths 1..N.
-                let synthetic = MatchOperation::Push {
-                    ctx_refs: contexts.clone(),
-                    pop_count,
-                };
-                if pop_count > 0 {
-                    // ST-observed divergence from plain `pop + set:`: on
-                    // `pop + embed:` the trigger match's text sees **neither**
-                    // `cur_context.meta_scope` nor `cur_context.meta_content_scope`.
-                    // Both are suppressed on match text, then never restored
-                    // — the embed replaces cur entirely. The probe lives at
-                    // the top of `v2_pop_embed_suppresses_cur_meta_scope_on_match`.
-                    //
-                    // Emit those Pops ourselves in the initial phase, then pass
-                    // a scope-stripped cur_context through to the recursive
-                    // Set-semantic logic so its `num_to_pop` in the non-initial
-                    // phase does not double-count these atoms (they are already
-                    // off the stack). clear_scopes, with_prototype, and other
-                    // fields are preserved on the stripped context — only the
-                    // meta-scope vectors differ.
-                    //
-                    // Observed divergence on `<jsp:declaration>`'s `>`:
-                    // syntect was producing
-                    //   [..., meta.tag.jsp.declaration.begin.html,
-                    //        meta.tag.jsp.declaration.begin.html,
-                    //        punctuation.definition.tag.end.html]
-                    // because the rule's explicit
-                    //   scope: meta.tag.jsp.declaration.begin.html
-                    //          punctuation.definition.tag.end.html
-                    // was re-adding the atom that ST drops through the embed.
-                    if initial {
-                        if !cur_context.meta_content_scope.is_empty() {
-                            ops.push((
-                                index,
-                                ScopeStackOp::Pop(cur_context.meta_content_scope.len()),
-                            ));
-                        }
-                        if !cur_context.meta_scope.is_empty() {
-                            ops.push((index, ScopeStackOp::Pop(cur_context.meta_scope.len())));
-                        }
-                    }
-                    let stripped = Context {
-                        meta_scope: Vec::new(),
-                        meta_content_scope: Vec::new(),
-                        ..cur_context.clone()
-                    };
-                    return self
-                        .push_meta_ops(initial, index, &stripped, &synthetic, syntax_set, ops);
-                }
-                return self.push_meta_ops(
-                    initial,
-                    index,
-                    cur_context,
-                    &synthetic,
-                    syntax_set,
-                    ops,
-                );
-            }
-            MatchOperation::None | MatchOperation::Fail(_) => (),
-            MatchOperation::Branch {
-                ref alternatives,
-                pop_count,
-                ..
-            } => {
-                // Branch acts like Push for meta ops purposes — `pop:N + branch:`
-                // is lookahead per ST, same as `pop:N + push:`. At exec time,
-                // Branch is transformed into a synthetic Push before calling
-                // push_meta_ops, so this arm is a safety fallback.
-                let synthetic = MatchOperation::Push {
-                    ctx_refs: alternatives.clone(),
-                    pop_count,
-                };
-                return self.push_meta_ops(
-                    initial,
-                    index,
-                    cur_context,
-                    &synthetic,
-                    syntax_set,
-                    ops,
-                );
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Returns true if the stack was changed
-    fn perform_op(
-        &mut self,
-        line: &str,
-        regions: &Region,
-        pat: &MatchPattern,
-        syntax_set: &SyntaxSet,
-    ) -> Result<bool, ParsingError> {
-        let (ctx_refs, old_proto_ids, is_embed) = match pat.operation {
-            MatchOperation::Push {
-                ref ctx_refs,
-                pop_count,
-            } => {
-                // `pop: N + push:` (pop_count > 0): pop N frames from the
-                // runtime stack before the push loop, mirroring Set's arm.
-                // The topmost popped frame's prototypes carry over so a
-                // `with_prototype` from the leaving context stays active on
-                // the new push (same convention as Set).
-                let old_proto_ids = if pop_count > 0 {
-                    let topmost = self.stack.pop().map(|s| s.prototypes);
-                    for _ in 1..pop_count {
-                        self.stack.pop();
-                    }
-                    topmost
-                } else {
-                    None
-                };
-                if pop_count > 0 {
-                    let final_len = self.stack.len() + ctx_refs.len();
-                    self.branch_points
-                        .retain(|bp| final_len > bp.stack_depth.saturating_sub(bp.pop_count));
-                    self.escape_stack.retain(|e| e.stack_depth < final_len);
-                }
-                (ctx_refs, old_proto_ids, false)
-            }
-            MatchOperation::Embed {
-                ref contexts,
-                pop_count,
-                ..
-            } => {
-                if pop_count > 0 {
-                    for _ in 0..pop_count {
-                        self.stack.pop();
-                    }
-                    let stack_len = self.stack.len();
-                    self.branch_points
-                        .retain(|bp| stack_len > bp.stack_depth.saturating_sub(bp.pop_count));
-                    self.escape_stack.retain(|e| e.stack_depth < stack_len);
-                }
-                (contexts, None, true)
-            }
-            MatchOperation::Set {
-                ref ctx_refs,
-                pop_count,
-            } => {
-                // a `with_prototype` stays active when the context is `set`
-                // until the context layer in the stack (where the `with_prototype`
-                // was initially applied) is popped off. With `pop: N + set:`
-                // (pop_count > 1), the topmost popped frame's prototypes are
-                // what carry forward onto the new push.
-                let pops = pop_count.max(1);
-                let old_proto_ids = self.stack.pop().map(|s| s.prototypes);
-                for _ in 1..pops {
-                    self.stack.pop();
-                }
-                // Prune branch_points / escape_stack against the *final* stack
-                // length (after the common push loop below).
-                //
-                // The retain predicate must mirror `handle_fail`'s validity
-                // check (`stack.len() > bp.stack_depth - bp.pop_count`),
-                // which subtracts the bp's own `pop_count`. Without that
-                // subtraction, a `pop: N + branch_point` whose synthetic
-                // Set has `pop_count: N` removes its own freshly-created
-                // bp here — `bp.stack_depth` snapshots the *pre-pop*
-                // depth, so `bp.stack_depth > final_len` even though the
-                // alt-0 frame lives on at `final_len`. Symptom in Java:
-                // the `branch_point: annotation-qualified-parameters`
-                // declared on `annotation-qualified-identifier-name`'s
-                // `pop: 2 + branch_point` was dropped at creation,
-                // making its later `(?=\S)` `fail` a no-op and leaking
-                // `meta.annotation.identifier.java meta.path.java` past
-                // every nested-annotation extends path.
-                let final_len = self.stack.len() + ctx_refs.len();
-                self.branch_points
-                    .retain(|bp| final_len > bp.stack_depth.saturating_sub(bp.pop_count));
-                self.escape_stack.retain(|e| e.stack_depth < final_len);
-                (ctx_refs, old_proto_ids, false)
-            }
-            MatchOperation::Pop(n) => {
-                for _ in 0..n {
-                    self.stack.pop();
-                }
-                // Invalidate branch points whose alt frame is no longer on
-                // the stack. Use the same threshold as `handle_fail`'s
-                // validity check — see the comment in the Set arm above.
-                let stack_len = self.stack.len();
-                self.branch_points
-                    .retain(|bp| stack_len > bp.stack_depth.saturating_sub(bp.pop_count));
-                // Remove escape entries whose stack_depth >= current stack
-                self.escape_stack.retain(|e| e.stack_depth < stack_len);
-                return Ok(true);
-            }
-            MatchOperation::None => return Ok(false),
-            MatchOperation::Branch { .. } | MatchOperation::Fail(_) => {
-                // Branch and Fail are handled in exec_pattern, not here
-                return Ok(false);
-            }
-        };
-
-        // Record stack depth before pushing (for Embed escape entry)
-        let stack_depth_before = self.stack.len();
-
-        for (i, r) in ctx_refs.iter().enumerate() {
-            let mut proto_ids = if i == 0 {
-                // it is only necessary to preserve the old prototypes
-                // at the first stack frame pushed
-                old_proto_ids.clone().unwrap_or_else(Vec::new)
-            } else {
-                Vec::new()
-            };
-            if i == ctx_refs.len() - 1 {
-                // if a with_prototype was specified, and multiple contexts were pushed,
-                // then the with_prototype applies only to the last context pushed, i.e.
-                // top most on the stack after all the contexts are pushed - this is also
-                // referred to as the "target" of the push by sublimehq - see
-                // https://forum.sublimetext.com/t/dev-build-3111/19240/17 for more info
-                if let Some(ref p) = pat.with_prototype {
-                    proto_ids.push(p.id()?);
-                }
-            }
-            let context_id = r.id()?;
-            let context = syntax_set.get_context(&context_id)?;
-            let captures = {
-                let mut uses_backrefs = context.uses_backrefs;
-                if !proto_ids.is_empty() {
-                    uses_backrefs = uses_backrefs
-                        || proto_ids
-                            .iter()
-                            .any(|id| syntax_set.get_context(id).unwrap().uses_backrefs);
-                }
-                if uses_backrefs {
-                    Some((regions.clone(), line.to_owned()))
-                } else {
-                    None
-                }
-            };
-            self.stack.push(StateLevel {
-                context: context_id,
-                prototypes: proto_ids,
-                captures,
-            });
-        }
-
-        // For Embed: push an EscapeEntry with the resolved escape regex
-        if is_embed {
-            if let MatchOperation::Embed { ref escape, .. } = pat.operation {
-                let resolved_regex = if escape.has_captures {
-                    // Resolve backrefs in escape regex using the triggering match's captures
-                    let new_regex_str =
-                        substitute_backrefs_in_regex(escape.escape_regex.regex_str(), |i| {
-                            regions.pos(i).map(|(s, e)| escape_str(&line[s..e]))
-                        });
-                    Regex::new(new_regex_str)
-                } else {
-                    escape.escape_regex.clone()
-                };
-                self.escape_stack.push(EscapeEntry {
-                    regex: resolved_regex,
-                    captures: escape.escape_captures.clone(),
-                    stack_depth: stack_depth_before,
-                });
-            }
-        }
-
-        Ok(true)
-    }
-
-    /// Execute an escape match: apply escape_captures, pop stack down to
-    /// the embed's stack_depth, and remove the escape entry.
-    fn exec_escape(
-        &mut self,
-        escape_idx: usize,
-        match_start: usize,
-        _match_end: usize,
-        regions: &Region,
-        syntax_set: &SyntaxSet,
-        ops: &mut Vec<(usize, ScopeStackOp)>,
-    ) -> Result<(), ParsingError> {
-        let entry = &self.escape_stack[escape_idx];
-        let target_depth = entry.stack_depth;
-        let escape_captures = entry.captures.clone();
-
-        // Drain orphan scope atoms left on the consumer's scope stack by
-        // a prior cross-line replay whose later same-line fails
-        // truncated the owning context out of `self.stack` — the Push
-        // was committed to `flushed_ops` and can't be unwound by
-        // `ops.truncate`, so we emit a balancing Pop here. Without this,
-        // e.g. LaTeX `\end{lstlisting}` leaves
-        // `meta.environment.verbatim.lstlisting.latex` on the stack
-        // because a speculative `meta.path.java` atom pushed inside the
-        // embedded Java shifts every subsequent Pop by one.
-        //
-        // `shadow` mirrors what the consumer will actually hold at this
-        // point: end-of-prior-line shadow + ops-so-far on the current
-        // line. `expected_depth` is what the consumer *should* have
-        // based on `self.stack`'s meta_scope / meta_content_scope
-        // contributions (with the v2 `embed_scope_replaces` mcs gating
-        // applied below, matching the pop loop).
-        let mut current_shadow = self.shadow.clone();
-        for (_, op) in ops.iter() {
-            let _ = current_shadow.apply(op);
-        }
-        let consumer_depth = current_shadow.as_slice().len();
-        let expected_depth: usize = {
-            let mut total = 0usize;
-            let mut prev_embed_scope_replaces = false;
-            for lvl in &self.stack {
-                let ctx = syntax_set.get_context(&lvl.context)?;
-                total += ctx.meta_scope.len();
-                if !prev_embed_scope_replaces {
-                    total += ctx.meta_content_scope.len();
-                }
-                prev_embed_scope_replaces = ctx.embed_scope_replaces;
-            }
-            total
-        };
-        if consumer_depth > expected_depth {
-            ops.push((
-                match_start,
-                ScopeStackOp::Pop(consumer_depth - expected_depth),
-            ));
-        }
-
-        // Pop all stack levels down to target_depth, emitting proper meta scope pops
-        while self.stack.len() > target_depth {
-            let level = &self.stack[self.stack.len() - 1];
-            let ctx = syntax_set.get_context(&level.context)?;
-
-            // Pop meta_content_scope.  If the context below has
-            // embed_scope_replaces (it's a v2 embed_scope wrapper), the top
-            // context is the embedded syntax's main — whose mcs was never
-            // pushed on the way in — so skip the pop here too.  Gating this
-            // on `current_syntax_version >= 2` would be wrong: the version
-            // is read from the top context (the embedded syntax), but
-            // embed_scope_replaces is set only by the v2 host syntax.  A v2
-            // host embedding a v1 grammar (e.g. Rails HTML embedding Ruby)
-            // would otherwise Pop a scope that was never pushed, misaligning
-            // every scope below until the escape closes.
-            if !ctx.meta_content_scope.is_empty() {
-                let skip = self.stack.len() >= 2
-                    && syntax_set
-                        .get_context(&self.stack[self.stack.len() - 2].context)
-                        .map(|c| c.embed_scope_replaces)
-                        .unwrap_or(false);
-                if !skip {
-                    ops.push((match_start, ScopeStackOp::Pop(ctx.meta_content_scope.len())));
-                }
-            }
-
-            // Pop meta_scope
-            if !ctx.meta_scope.is_empty() {
-                ops.push((match_start, ScopeStackOp::Pop(ctx.meta_scope.len())));
-            }
-
-            // Restore cleared scopes
-            if ctx.clear_scopes.is_some() {
-                ops.push((match_start, ScopeStackOp::Restore));
-            }
-
-            self.stack.pop();
-        }
-
-        // Apply escape_captures scopes
-        if let Some(ref capture_map) = escape_captures {
-            let mut map: Vec<((usize, i32), ScopeStackOp)> = Vec::new();
-            for &(cap_index, ref scopes) in capture_map.iter() {
-                if let Some((cap_start, cap_end)) = regions.pos(cap_index) {
-                    if cap_start == cap_end {
-                        continue;
-                    }
-                    for scope in scopes.iter() {
-                        map.push((
-                            (cap_start, -((cap_end - cap_start) as i32)),
-                            ScopeStackOp::Push(*scope),
-                        ));
-                    }
-                    map.push(((cap_end, i32::MIN), ScopeStackOp::Pop(scopes.len())));
-                }
-            }
-            map.sort_by(|a, b| a.0.cmp(&b.0));
-            for ((index, _), op) in map.into_iter() {
-                ops.push((index, op));
-            }
-        }
-
-        // Remove this escape entry and any inner (later) escape entries
-        self.escape_stack.truncate(escape_idx);
-
-        // Invalidate branch points whose alt frame is no longer on the
-        // stack. This mirrors the `alt frame still present` predicate used
-        // at the other five BP-prune sites (handle_fail's late guard plus
-        // the Push/Set/Embed/Pop retain calls): subtracting the bp's own
-        // pop_count is necessary so a `pop: N + branch_point` whose
-        // snapshot captures the pre-pop depth doesn't false-prune itself.
-        let stack_len = self.stack.len();
-        self.branch_points
-            .retain(|bp| stack_len > bp.stack_depth.saturating_sub(bp.pop_count));
-
-        Ok(())
-    }
-}
-
-/// Escape a string for use in regex substitution (re-export for use in escape resolution).
-fn escape_str(s: &str) -> String {
-    escape(s)
 }
 
 #[cfg(feature = "yaml-load")]
